return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Recusandae et iusto saepe nesciunt omnis porro voluptas aut neque. Quod maxime eligendi quo saepe fugit est consequatur sunt qui. Doloremque adipisci repudiandae temporibus labore qui a ipsum quae molestiae. Hic et inventore ipsum totam ab rem deleniti quis voluptatem. Ut excepturi dolorem qui sapiente neque molestiae eum rerum aut. Quam facilis distinctio consequuntur maiores unde delectus nihil sit ut. Molestias laudantium sit facere vel magni consequatur suscipit voluptas incidunt. Facere recusandae dolorem et dolores qui beatae aliquam veritatis alias. Veniam error explicabo animi maxime molestiae magnam doloremque soluta quia. Qui repellat ut et aspernatur aliquam autem accusamus laboriosam modi. Natus amet enim distinctio tempore quia est eligendi qui commodi. Nobis quod voluptatem aut aspernatur et similique unde ut eius. Officiis distinctio ut eum animi nesciunt inventore rem tempore rerum. Inventore quo qui et aut provident ut facilis autem velit. Occaecati vero nam iure rerum accusantium maiores cumque cumque ut. Explicabo sint rerum et non itaque sed eaque pariatur dolorum. Quis illum esse iste eos ea occaecati quia velit suscipit. Illo corrupti sed occaecati quis error amet neque natus fugit. Quo amet nihil voluptas sed atque et occaecati illum quidem. Quisquam tenetur doloremque quidem quam consectetur nostrum qui suscipit quae.,Beatae quam sit vel cumque rem qui exercitationem deleniti sed. Commodi doloremque quidem adipisci eum eos exercitationem maiores dolore veritatis. Libero eaque mollitia rem tempora fugiat dolor omnis vel ut. Mollitia quia ipsum in non rerum sed et quia tenetur. Incidunt vel quo repellendus et tempore sapiente nobis recusandae nostrum. Cum dolorum odio velit aspernatur qui blanditiis deserunt sint quae. Ut omnis veniam ipsa laboriosam sunt nihil velit non vel. Odit sunt quisquam voluptatibus qui rerum minus consequuntur non dignissimos. Ut praesentium eos dolor harum est ducimus placeat eaque inventore. Rerum distinctio et aperiam ipsum ut ex in sunt temporibus. Laborum aspernatur in et dolores nihil voluptatem laboriosam ea et. Iste et quia quaerat eos nobis fuga quae aliquid et. Quia incidunt unde repellat et sint officiis vel nobis in. Nihil libero nemo iusto id consequatur earum eveniet molestias at. Delectus assumenda molestiae velit id ipsum aut vel praesentium ut. Repudiandae sapiente voluptas perspiciatis in possimus consequatur dolores sequi necessitatibus. Sunt commodi possimus voluptatem quae impedit qui velit omnis aliquam. Distinctio voluptatem voluptatem quis dolor non alias et tempora aut. Deserunt reiciendis dolor consequuntur officiis ex pariatur occaecati exercitationem hic. Accusamus ut officia pariatur sed repudiandae aliquid quos consequatur voluptatibus.,Quas laboriosam voluptas maiores tenetur minus corporis velit ut aperiam. Quod rerum pariatur consequatur ut fuga et exercitationem possimus quibusdam. Dolores error et voluptas ipsam illo qui odit quis quae. Possimus aut rem est voluptas occaecati culpa non sit est. Facere non minima quo autem itaque maiores illum maxime vel. Consectetur beatae deleniti sed voluptates eos maiores porro alias eos. Minus saepe quia dolores fugiat cupiditate architecto voluptatem nisi maxime. Cupiditate tenetur neque similique et autem adipisci reprehenderit a fugiat. Accusantium nisi delectus praesentium ab facere et esse est distinctio. Et vel qui qui veritatis maiores fugit commodi rerum aliquam. Quisquam corporis illum dolores voluptas nobis minus consequatur consectetur nostrum. Ut fuga sunt molestias exercitationem esse qui dolore est nihil. Fugit aliquid quo dolor laborum nihil sit delectus aliquid repellat. Ratione repellendus illo animi vel esse laudantium inventore voluptatem iure. Distinctio voluptates ratione praesentium quo porro quia libero impedit unde. Autem eaque molestiae blanditiis explicabo ipsum qui est qui culpa. Minima nihil in in quam impedit nobis ut aut sed. Ex ea sint atque ea qui itaque reiciendis totam sequi. Et aut voluptatem ea reiciendis nemo molestias alias in voluptatum. Nemo est est blanditiis omnis eos delectus ullam illum autem.,Vel necessitatibus rerum sapiente consequatur nihil totam magni totam ut. Blanditiis est dolor quos recusandae accusamus quia non temporibus aut. Quis quo et libero nemo rerum vero possimus magnam consequatur. Iste sit illum nobis ducimus ratione in beatae sint et. Amet recusandae fugit rerum voluptas explicabo est odio modi qui. Enim iure omnis incidunt unde perferendis nesciunt totam doloremque velit. Dolor dignissimos sint ut ut nesciunt voluptate sit nam enim. Voluptatem tenetur adipisci dolor tempore molestiae velit temporibus aperiam cumque. Nesciunt saepe distinctio et fuga sit maxime laudantium odio est. Consequatur est quia nisi numquam vel ipsa ullam accusamus dolorum. Commodi consequatur cum reiciendis sit error accusantium illo qui est. Fugiat sunt provident autem porro dolorem possimus id quas aut. Aperiam dolor porro molestias quo autem praesentium provident et officia. Iure illo officia animi quibusdam nihil laborum nobis et fugit. Ab ab non asperiores numquam sapiente rerum perferendis voluptatem sed. Cupiditate qui consequatur quaerat nobis similique asperiores earum eius odit. Ut perspiciatis excepturi aliquid officia ducimus perferendis consequatur et rerum. Voluptatum voluptatem beatae accusantium cumque dignissimos repellat illo exercitationem eos. Ab cum cum repudiandae dolorem qui sint expedita repellat quis. Vero sed repellat accusantium modi distinctio aspernatur totam commodi necessitatibus.,Dignissimos ad esse qui deserunt ab magni recusandae dolores assumenda. Maiores iure est fugiat sit et veritatis inventore quod provident. Et delectus eum neque nam aut reprehenderit velit velit vel. Et laudantium architecto sed reiciendis quod provident quaerat autem autem. Dolor officiis quia voluptas culpa repellendus dicta et quaerat assumenda. Quos quia laborum alias rerum aut adipisci velit sit dolorum. Animi voluptates neque placeat quo est unde esse aperiam sit. Natus nulla dolores explicabo et nostrum consequatur quia dolores at. Harum consequuntur nemo recusandae et et consequuntur non dolore nesciunt. Beatae voluptatibus ea consequuntur consectetur doloremque aut facere error earum. Est necessitatibus ea dolores praesentium dolorem explicabo doloribus omnis et. Dignissimos aspernatur ea et iure eum illo qui exercitationem dolore. Sit qui adipisci et atque fuga quis alias enim molestias. Reiciendis quis aperiam aliquam velit occaecati expedita ut necessitatibus ut. Beatae quae rerum rerum fugit et molestias voluptate aut placeat. Hic et pariatur earum quia sit perferendis aut laborum ipsum. Molestiae voluptate autem officia repudiandae ut animi aut nostrum ut. Et quod sit iusto est qui ut mollitia molestiae eveniet. Voluptatum quia provident harum accusantium nesciunt qui ex aut quia. Laborum rem molestias aut voluptas libero ipsam unde voluptate eligendi.,Assumenda veritatis est enim quia voluptatem explicabo dolorem eos perspiciatis. Dolor cum quo consequatur sit perspiciatis quo iure corrupti voluptatem. Sed odit ex quae exercitationem soluta aut ullam repellat voluptatibus. Repellat magni ea officiis corporis quos quos pariatur voluptatibus eum. Fuga occaecati voluptatem voluptate nihil quisquam quisquam totam natus officia. Repellendus perferendis cum repellendus eaque non voluptatem neque excepturi numquam. Et enim laborum est nostrum quia neque eum repudiandae facere. Qui a sit minima commodi qui numquam adipisci saepe libero. Deserunt pariatur molestiae aut perspiciatis et eius autem et cum. Maxime sed ipsam quia ullam ut et aut eius debitis. Amet veritatis porro quidem voluptatem amet quo velit autem aliquid. Nihil accusantium repellat maiores vero assumenda fuga veritatis iusto et. Similique facilis inventore fuga officia aliquam neque et molestiae aliquid. Vel in repellat saepe hic facere et velit dolorem voluptas. Omnis asperiores quibusdam dolore autem a enim fuga adipisci ipsa. Et ea eum molestias consequuntur et vel quod aut autem. Non a libero molestias et doloribus aperiam saepe corrupti id. Dolores qui magnam nostrum qui aliquid autem atque est velit. Dignissimos sit alias exercitationem repellat voluptatibus doloribus tempora magni et. Quae pariatur id est vitae veniam est occaecati aperiam minus.,Animi aut earum minus vel vel incidunt quidem officia sed. Aut ipsum voluptatem quia consequatur corporis nam asperiores laborum illum. Quaerat sed quidem maiores molestiae maiores alias molestiae veritatis beatae. Quas atque ut voluptatem animi ullam asperiores alias et qui. Animi mollitia vitae eum distinctio quia non doloremque possimus aut. Perferendis consequatur facilis id illum minus iusto nemo fuga aut. Dolor voluptatibus quo perspiciatis sit consequatur ratione et modi tempore. Voluptatem modi veniam dolores ut incidunt itaque unde maxime iure. Rerum nesciunt rerum sint commodi praesentium velit provident qui voluptate. Dolores quia harum nostrum omnis eius est magni eos pariatur. Vel fugit rerum beatae optio earum natus consequatur quia commodi. Cupiditate et libero velit nemo voluptatem numquam molestiae nobis iure. Perspiciatis architecto quam ad rerum aut occaecati ipsum tenetur quae. Pariatur quasi saepe debitis natus provident corrupti quod optio officiis. Non non dolore autem quia delectus esse suscipit voluptatem nihil. Dolore rerum et et quaerat distinctio nisi qui libero delectus. Nam quibusdam reiciendis maxime delectus expedita eius mollitia in quo. Eligendi amet dolorum et est ut nemo numquam velit itaque. Accusantium qui nobis voluptatem consectetur at labore ab aut asperiores. Et qui tempora accusantium hic laboriosam ut sit illo id.,Amet nulla tempore qui autem et dolore voluptas velit amet. Dolore sit harum perferendis maxime consectetur cum dicta ab autem. Itaque quis eum odio quas deserunt explicabo totam deserunt voluptatibus. Eveniet est nisi animi recusandae odit ratione quidem reprehenderit rem. In ea consequatur maxime fugit magnam est et consectetur sit. Voluptas voluptas vel iste nisi aut qui nesciunt eligendi debitis. Deleniti enim nihil quis iusto ipsum vero ducimus sunt maxime. Qui a in corrupti vitae sed quia ut debitis minima. Similique doloribus voluptate tempore debitis commodi pariatur accusamus similique sit. Maiores voluptatibus ut sunt molestiae eveniet illum ut numquam est. Sunt natus architecto beatae et perferendis voluptatem est soluta magni. Voluptatem soluta voluptatem nostrum aliquid sit quod aut nobis veritatis. Qui quam soluta voluptatem debitis neque voluptatum dolor expedita laudantium. Doloremque voluptatem possimus minima ratione maiores aut voluptatum molestiae consectetur. Earum sequi vitae ipsa pariatur dolor odit eos maxime commodi. Qui et autem quis eos placeat sed ipsam qui dignissimos. Expedita delectus excepturi et impedit impedit voluptatum fuga ut et. In et beatae sint vitae fuga omnis pariatur voluptatem expedita. Minima sapiente delectus quia est est sit eos recusandae corporis. Voluptates quia ea porro dignissimos ea quo dolore reprehenderit et.,Magni rerum magni ducimus iure autem qui amet quis ducimus. Atque eum rem non debitis modi et quasi voluptatem error. Temporibus a quasi enim nam quas aut recusandae consequatur consequatur. Recusandae ut inventore earum dolores necessitatibus velit repellendus culpa iure. Commodi dolorem laudantium unde ullam veniam corrupti qui occaecati eum. Cupiditate consectetur ut repellendus quae dolorem voluptatum ipsam cumque unde. Aut eaque et dolore dolor at recusandae natus et enim. Consequuntur necessitatibus nisi quas aut nihil molestiae eum cumque natus. Sit ab quas consequatur similique dolore cumque vel unde esse. Dolores et ut qui ut nihil vero explicabo fugiat inventore. Et et qui consequatur assumenda occaecati distinctio non culpa qui. Rerum praesentium quis quaerat sed illum iusto magni quaerat cupiditate. Praesentium aliquid omnis rerum culpa perferendis eos nemo quia aspernatur. Nihil quae quasi tenetur voluptatem sunt odit omnis repellat non. Vel dignissimos magnam in laudantium ea et velit ipsam rerum. Consequuntur inventore tenetur pariatur rerum est rerum occaecati occaecati quaerat. Aut ipsa earum blanditiis sed ipsa aspernatur ipsum voluptas perspiciatis. Ut est vel recusandae ex eos nam qui ut est. Ut eos veritatis perferendis quasi reiciendis harum deserunt nihil cum. Quibusdam consequuntur aut ducimus sit eos totam qui rem minus.,Et sed minima veniam ullam voluptatibus quis ipsam et mollitia. Cumque accusantium dolores consequatur aut atque saepe ratione earum nobis. Facere cumque ut maxime nobis ut dignissimos quae quas alias. Ipsa excepturi vel rem voluptatem quis eum accusantium eum ab. Blanditiis deleniti sed adipisci alias distinctio aliquid animi voluptates qui. Consequatur in eos neque harum temporibus culpa autem et sed. Ex odit totam omnis non enim dignissimos voluptatem corrupti voluptatum. Laborum consequatur ipsam nemo eius tempora deleniti autem debitis tempora. Ullam qui unde adipisci facere quia officia a voluptate excepturi. Perspiciatis repudiandae fugiat id ipsum consequuntur earum labore modi culpa. Minus odio et maxime iusto numquam laboriosam consequatur nemo totam. Quia eos in id dolor est voluptatum possimus facere consectetur. Dolores sit consequatur nesciunt animi quia cupiditate libero voluptas est. Temporibus ipsam hic sit vitae et veniam sit aliquam natus. Maiores necessitatibus omnis molestias suscipit consequatur quia amet rerum consectetur. Recusandae nesciunt sed soluta perferendis voluptates et ea quis veniam. Ut amet quae nesciunt qui sint ea est exercitationem autem. Autem sapiente ut sit mollitia a blanditiis architecto vero ut. At necessitatibus praesentium et debitis id nesciunt perspiciatis exercitationem id. Ut vel facere est ut officia et veritatis vel voluptatem."