read excel sheet by name to json

This is a function to read out the content out of a excel sheet to json array

read_sheet_by_name(file_path, sheet_name)

Examples

read_sheet_by_name("example_input.xlsx","Test1")
#> [1] "[{\"_row\":\"Mazda RX4\",\"am\":\"1\",\"carb\":\"4\",\"cyl\":\"6\",\"disp\":\"160\",\"drat\":\"3.9\",\"gear\":\"4\",\"hp\":\"110\",\"mpg\":\"21\",\"qsec\":\"16.46\",\"vs\":\"0\",\"wt\":\"2.62\"},{\"_row\":\"Mazda RX4 Wag\",\"am\":\"1\",\"carb\":\"4\",\"cyl\":\"6\",\"disp\":\"160\",\"drat\":\"3.9\",\"gear\":\"4\",\"hp\":\"110\",\"mpg\":\"21\",\"qsec\":\"17.02\",\"vs\":\"0\",\"wt\":\"2.875\"},{\"_row\":\"Datsun 710\",\"am\":\"1\",\"carb\":\"1\",\"cyl\":\"4\",\"disp\":\"108\",\"drat\":\"3.85\",\"gear\":\"4\",\"hp\":\"93\",\"mpg\":\"22.8\",\"qsec\":\"18.61\",\"vs\":\"1\",\"wt\":\"2.32\"},{\"_row\":\"Hornet 4 Drive\",\"am\":\"0\",\"carb\":\"1\",\"cyl\":\"6\",\"disp\":\"258\",\"drat\":\"3.08\",\"gear\":\"3\",\"hp\":\"110\",\"mpg\":\"21.4\",\"qsec\":\"19.44\",\"vs\":\"1\",\"wt\":\"3.215\"},{\"_row\":\"Hornet Sportabout\",\"am\":\"0\",\"carb\":\"2\",\"cyl\":\"8\",\"disp\":\"360\",\"drat\":\"3.15\",\"gear\":\"3\",\"hp\":\"175\",\"mpg\":\"18.7\",\"qsec\":\"17.02\",\"vs\":\"0\",\"wt\":\"3.44\"},{\"_row\":\"Valiant\",\"am\":\"0\",\"carb\":\"1\",\"cyl\":\"6\",\"disp\":\"225\",\"drat\":\"2.76\",\"gear\":\"3\",\"hp\":\"105\",\"mpg\":\"18.1\",\"qsec\":\"20.22\",\"vs\":\"1\",\"wt\":\"3.46\"},{\"_row\":\"Duster 360\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"8\",\"disp\":\"360\",\"drat\":\"3.21\",\"gear\":\"3\",\"hp\":\"245\",\"mpg\":\"14.3\",\"qsec\":\"15.84\",\"vs\":\"0\",\"wt\":\"3.57\"},{\"_row\":\"Merc 240D\",\"am\":\"0\",\"carb\":\"2\",\"cyl\":\"4\",\"disp\":\"146.7\",\"drat\":\"3.69\",\"gear\":\"4\",\"hp\":\"62\",\"mpg\":\"24.4\",\"qsec\":\"20\",\"vs\":\"1\",\"wt\":\"3.19\"},{\"_row\":\"Merc 230\",\"am\":\"0\",\"carb\":\"2\",\"cyl\":\"4\",\"disp\":\"140.8\",\"drat\":\"3.92\",\"gear\":\"4\",\"hp\":\"95\",\"mpg\":\"22.8\",\"qsec\":\"22.9\",\"vs\":\"1\",\"wt\":\"3.15\"},{\"_row\":\"Merc 280\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"6\",\"disp\":\"167.6\",\"drat\":\"3.92\",\"gear\":\"4\",\"hp\":\"123\",\"mpg\":\"19.2\",\"qsec\":\"18.3\",\"vs\":\"1\",\"wt\":\"3.44\"},{\"_row\":\"Merc 280C\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"6\",\"disp\":\"167.6\",\"drat\":\"3.92\",\"gear\":\"4\",\"hp\":\"123\",\"mpg\":\"17.8\",\"qsec\":\"18.9\",\"vs\":\"1\",\"wt\":\"3.44\"},{\"_row\":\"Merc 450SE\",\"am\":\"0\",\"carb\":\"3\",\"cyl\":\"8\",\"disp\":\"275.8\",\"drat\":\"3.07\",\"gear\":\"3\",\"hp\":\"180\",\"mpg\":\"16.4\",\"qsec\":\"17.4\",\"vs\":\"0\",\"wt\":\"4.07\"},{\"_row\":\"Merc 450SL\",\"am\":\"0\",\"carb\":\"3\",\"cyl\":\"8\",\"disp\":\"275.8\",\"drat\":\"3.07\",\"gear\":\"3\",\"hp\":\"180\",\"mpg\":\"17.3\",\"qsec\":\"17.6\",\"vs\":\"0\",\"wt\":\"3.73\"},{\"_row\":\"Merc 450SLC\",\"am\":\"0\",\"carb\":\"3\",\"cyl\":\"8\",\"disp\":\"275.8\",\"drat\":\"3.07\",\"gear\":\"3\",\"hp\":\"180\",\"mpg\":\"15.2\",\"qsec\":\"18\",\"vs\":\"0\",\"wt\":\"3.78\"},{\"_row\":\"Cadillac Fleetwood\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"8\",\"disp\":\"472\",\"drat\":\"2.93\",\"gear\":\"3\",\"hp\":\"205\",\"mpg\":\"10.4\",\"qsec\":\"17.98\",\"vs\":\"0\",\"wt\":\"5.25\"},{\"_row\":\"Lincoln Continental\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"8\",\"disp\":\"460\",\"drat\":\"3\",\"gear\":\"3\",\"hp\":\"215\",\"mpg\":\"10.4\",\"qsec\":\"17.82\",\"vs\":\"0\",\"wt\":\"5.424\"},{\"_row\":\"Chrysler Imperial\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"8\",\"disp\":\"440\",\"drat\":\"3.23\",\"gear\":\"3\",\"hp\":\"230\",\"mpg\":\"14.7\",\"qsec\":\"17.42\",\"vs\":\"0\",\"wt\":\"5.345\"},{\"_row\":\"Fiat 128\",\"am\":\"1\",\"carb\":\"1\",\"cyl\":\"4\",\"disp\":\"78.7\",\"drat\":\"4.08\",\"gear\":\"4\",\"hp\":\"66\",\"mpg\":\"32.4\",\"qsec\":\"19.47\",\"vs\":\"1\",\"wt\":\"2.2\"},{\"_row\":\"Honda Civic\",\"am\":\"1\",\"carb\":\"2\",\"cyl\":\"4\",\"disp\":\"75.7\",\"drat\":\"4.93\",\"gear\":\"4\",\"hp\":\"52\",\"mpg\":\"30.4\",\"qsec\":\"18.52\",\"vs\":\"1\",\"wt\":\"1.615\"},{\"_row\":\"Toyota Corolla\",\"am\":\"1\",\"carb\":\"1\",\"cyl\":\"4\",\"disp\":\"71.1\",\"drat\":\"4.22\",\"gear\":\"4\",\"hp\":\"65\",\"mpg\":\"33.9\",\"qsec\":\"19.9\",\"vs\":\"1\",\"wt\":\"1.835\"},{\"_row\":\"Toyota Corona\",\"am\":\"0\",\"carb\":\"1\",\"cyl\":\"4\",\"disp\":\"120.1\",\"drat\":\"3.7\",\"gear\":\"3\",\"hp\":\"97\",\"mpg\":\"21.5\",\"qsec\":\"20.01\",\"vs\":\"1\",\"wt\":\"2.465\"},{\"_row\":\"Dodge Challenger\",\"am\":\"0\",\"carb\":\"2\",\"cyl\":\"8\",\"disp\":\"318\",\"drat\":\"2.76\",\"gear\":\"3\",\"hp\":\"150\",\"mpg\":\"15.5\",\"qsec\":\"16.87\",\"vs\":\"0\",\"wt\":\"3.52\"},{\"_row\":\"AMC Javelin\",\"am\":\"0\",\"carb\":\"2\",\"cyl\":\"8\",\"disp\":\"304\",\"drat\":\"3.15\",\"gear\":\"3\",\"hp\":\"150\",\"mpg\":\"15.2\",\"qsec\":\"17.3\",\"vs\":\"0\",\"wt\":\"3.435\"},{\"_row\":\"Camaro Z28\",\"am\":\"0\",\"carb\":\"4\",\"cyl\":\"8\",\"disp\":\"350\",\"drat\":\"3.73\",\"gear\":\"3\",\"hp\":\"245\",\"mpg\":\"13.3\",\"qsec\":\"15.41\",\"vs\":\"0\",\"wt\":\"3.84\"},{\"_row\":\"Pontiac Firebird\",\"am\":\"0\",\"carb\":\"2\",\"cyl\":\"8\",\"disp\":\"400\",\"drat\":\"3.08\",\"gear\":\"3\",\"hp\":\"175\",\"mpg\":\"19.2\",\"qsec\":\"17.05\",\"vs\":\"0\",\"wt\":\"3.845\"},{\"_row\":\"Fiat X1-9\",\"am\":\"1\",\"carb\":\"1\",\"cyl\":\"4\",\"disp\":\"79\",\"drat\":\"4.08\",\"gear\":\"4\",\"hp\":\"66\",\"mpg\":\"27.3\",\"qsec\":\"18.9\",\"vs\":\"1\",\"wt\":\"1.935\"},{\"_row\":\"Porsche 914-2\",\"am\":\"1\",\"carb\":\"2\",\"cyl\":\"4\",\"disp\":\"120.3\",\"drat\":\"4.43\",\"gear\":\"5\",\"hp\":\"91\",\"mpg\":\"26\",\"qsec\":\"16.7\",\"vs\":\"0\",\"wt\":\"2.14\"},{\"_row\":\"Lotus Europa\",\"am\":\"1\",\"carb\":\"2\",\"cyl\":\"4\",\"disp\":\"95.1\",\"drat\":\"3.77\",\"gear\":\"5\",\"hp\":\"113\",\"mpg\":\"30.4\",\"qsec\":\"16.9\",\"vs\":\"1\",\"wt\":\"1.513\"},{\"_row\":\"Ford Pantera L\",\"am\":\"1\",\"carb\":\"4\",\"cyl\":\"8\",\"disp\":\"351\",\"drat\":\"4.22\",\"gear\":\"5\",\"hp\":\"264\",\"mpg\":\"15.8\",\"qsec\":\"14.5\",\"vs\":\"0\",\"wt\":\"3.17\"},{\"_row\":\"Ferrari Dino\",\"am\":\"1\",\"carb\":\"6\",\"cyl\":\"6\",\"disp\":\"145\",\"drat\":\"3.62\",\"gear\":\"5\",\"hp\":\"175\",\"mpg\":\"19.7\",\"qsec\":\"15.5\",\"vs\":\"0\",\"wt\":\"2.77\"},{\"_row\":\"Maserati Bora\",\"am\":\"1\",\"carb\":\"8\",\"cyl\":\"8\",\"disp\":\"301\",\"drat\":\"3.54\",\"gear\":\"5\",\"hp\":\"335\",\"mpg\":\"15\",\"qsec\":\"14.6\",\"vs\":\"0\",\"wt\":\"3.57\"},{\"_row\":\"Volvo 142E\",\"am\":\"1\",\"carb\":\"2\",\"cyl\":\"4\",\"disp\":\"121\",\"drat\":\"4.11\",\"gear\":\"4\",\"hp\":\"109\",\"mpg\":\"21.4\",\"qsec\":\"18.6\",\"vs\":\"1\",\"wt\":\"2.78\"}]"