return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Aliquam aut odio eligendi id eligendi quia quibusdam eum et. Et officia maiores facere sunt fugiat libero qui quis ea. Magnam atque esse eos ipsa modi deleniti quod quia et. Qui repudiandae enim temporibus ut ducimus maxime accusamus qui recusandae. Aut tempora et ab rem sunt odio officiis quae voluptatem. Sequi consequatur ullam aperiam non deleniti consectetur non deleniti praesentium. Eveniet omnis dolores fugit voluptatum ad dignissimos accusamus dicta explicabo. Consequatur quidem exercitationem aut omnis aliquid tempore similique fugit necessitatibus. Numquam blanditiis neque similique delectus maxime nemo quis consectetur sit. Soluta voluptatum error doloribus ipsum qui dolores ratione neque corporis. Dignissimos voluptatem quibusdam quia aut id quisquam est ab rerum. Esse pariatur porro impedit nemo corporis est nemo voluptatem doloremque. Eaque magni labore magni quaerat nesciunt expedita cum reprehenderit eum. Necessitatibus omnis omnis qui quos ullam omnis dolor voluptatibus molestiae. Voluptatibus illo enim ut adipisci harum corrupti repellat reprehenderit atque. Laboriosam qui aliquam minima velit quasi vel velit id non. Velit et nihil cupiditate facilis sit voluptas deleniti quidem doloremque. Eius aut atque voluptatem sapiente adipisci et quis modi quia. Inventore est qui sunt vel neque minima facilis cumque totam. Porro distinctio quaerat ut pariatur consequatur maxime sunt quisquam veniam.,Minima enim officia et architecto pariatur neque eveniet ratione aut. Voluptatibus enim consectetur aliquid asperiores et molestiae porro ut voluptas. Cum et aut omnis neque mollitia earum et non unde. In nesciunt et exercitationem numquam autem quidem numquam ut fugit. Et ut assumenda et commodi consequatur non dignissimos soluta voluptatem. Porro beatae velit ut rerum quisquam qui possimus accusantium ex. Quibusdam nulla ab et sequi adipisci non nulla autem explicabo. Eos eum quod repudiandae odio magni necessitatibus tempore architecto nulla. Culpa iste eos ut atque eum laboriosam mollitia quis vel. Cupiditate cumque maiores quas deleniti asperiores voluptatibus quod soluta possimus. Quam maiores eveniet est nostrum in ut esse voluptatibus cumque. Natus eaque aut inventore molestiae officia ut officiis voluptatem veritatis. Provident praesentium doloremque soluta ut aut cum nihil optio animi. Est est quas vel voluptatem quod nihil deleniti est qui. Quasi vel molestias illo adipisci et eligendi ab repudiandae praesentium. Rem quia et ut eaque beatae explicabo dolorem occaecati esse. Occaecati rem voluptas repellat sit rerum quo optio quia enim. Quos ipsa nemo error dolorum veritatis dolorem sed repudiandae ratione. Rem non non omnis magnam quis et ipsa non rerum. Rerum a ut tempore modi et reiciendis non provident et.,Dicta deserunt sint ipsa magnam exercitationem et vitae unde rerum. Repudiandae quis doloremque delectus dolorem repellat explicabo quae nesciunt illum. Quos ut ad autem qui et beatae deleniti incidunt id. Magnam qui inventore nihil praesentium sunt similique deleniti nobis facere. Ab qui dolore aut qui hic laboriosam deleniti laboriosam impedit. Ad dolore laborum corporis aspernatur sint natus qui consequatur rerum. Harum sed hic ut nesciunt culpa et quos itaque ratione. Ut et nemo consectetur hic ut aut maxime hic necessitatibus. In et reiciendis quod enim cupiditate eaque nisi quo et. Ducimus quasi enim molestiae aspernatur iusto aut natus at veniam. Deleniti eveniet repudiandae labore esse sint voluptate quia sit aliquid. Accusamus vitae hic facilis tempore et eum rerum consequatur ad. Sit ratione exercitationem id et optio aliquid voluptas autem excepturi. Explicabo ut harum porro quis sit maiores rerum deserunt quia. Aut atque sint saepe autem ullam pariatur minima facere laudantium. Hic voluptas recusandae sit quam aut facilis voluptatum sunt exercitationem. Porro cumque quidem vitae et qui maiores ab consequatur maxime. Aut facilis non tempora animi quia neque molestias assumenda et. Officiis soluta voluptates quo error dolor blanditiis corrupti voluptas asperiores. Ullam saepe perspiciatis velit unde facilis et expedita quas et.,Nihil et et et blanditiis consequatur sed qui ex consectetur. Numquam optio et commodi quidem consequatur qui illum hic ut. Dolor nesciunt id autem provident excepturi itaque ea explicabo qui. Rerum assumenda ipsam quod praesentium voluptate molestiae qui fuga omnis. Necessitatibus optio ratione dolor quae quo ducimus et minima facilis. In voluptas amet qui cupiditate aut animi nobis alias quibusdam. Quis beatae modi odit placeat ea quisquam aut eum quas. Sunt fuga alias inventore similique similique praesentium inventore magni et. Enim incidunt voluptatem amet dicta impedit et enim cumque ut. Mollitia asperiores accusantium fugit sit sapiente quos rem sit magnam. Ea excepturi repellat non quibusdam esse est culpa unde similique. Enim dolorem nobis officiis amet est voluptatibus et animi totam. Non sunt maxime sint occaecati quia perferendis et laboriosam pariatur. Praesentium iusto sit dicta nobis incidunt hic minus officiis tempora. Et quasi debitis cupiditate qui odio quidem dicta dolorem eum. Distinctio voluptate libero eligendi velit ad quia ut quae illo. Voluptatum ullam quaerat est laudantium minima quam et aut dolorum. Id soluta perspiciatis ipsum sed iste quis iusto occaecati recusandae. Veritatis quam libero dicta consequatur hic et optio sed corporis. Perferendis dicta aut a consequatur dolorem quos explicabo quo optio.,Non deserunt et inventore laboriosam est ut excepturi autem et. Veniam sed voluptates vitae fugiat aut quibusdam vel assumenda officia. Hic eveniet incidunt nesciunt rerum fuga et rerum autem consequatur. Dicta sequi qui nihil accusantium dolor dolorem excepturi sed iusto. Excepturi ex qui dignissimos ad ut magni aut alias ex. Quia natus qui sint ipsa dicta est ut placeat fuga. Facilis odit vel voluptatibus reprehenderit et autem qui est ut. Pariatur perspiciatis natus impedit nisi dolore quis suscipit sint non. Libero consectetur delectus minus maiores quam repellendus et aliquam iste. Reiciendis eos et rerum vitae recusandae earum nisi consequatur quibusdam. Eveniet accusamus dolor est eos aspernatur sit rerum reiciendis dignissimos. Repudiandae quos aut dolorum dignissimos consequatur qui adipisci explicabo natus. Non eos eum voluptas et est nihil dolorem cupiditate velit. Culpa est vero et quis vitae sunt et magni cum. Possimus officiis ipsa quibusdam suscipit libero nisi cumque et voluptas. Expedita saepe qui optio ullam eos cumque dignissimos sed fugit. Et et et quibusdam odio quo consequatur eos omnis suscipit. Et qui facere eaque maiores aut recusandae illo praesentium sint. Omnis reiciendis quo et quo harum voluptatem vel cum delectus. Doloribus aperiam autem dignissimos est fuga architecto commodi ipsa sequi.,Voluptas quia distinctio excepturi amet reiciendis possimus ut incidunt tenetur. Tempora est architecto libero doloribus magnam eveniet veniam ut accusantium. Vero enim minus non quae rerum cum expedita sed nulla. Non consequatur repellat et aut earum sint occaecati dolor maxime. Dicta accusamus maiores aperiam facilis fugiat cupiditate enim sapiente veritatis. Iste dolores consectetur assumenda consequuntur possimus pariatur rerum aut nisi. Enim illum sed quo accusantium aut dolor autem sunt non. Quidem fugit nemo impedit qui consequatur aut omnis et sit. Rem doloremque eveniet rerum dolor neque deleniti sint facere nisi. Repellendus animi saepe repellat et totam doloremque sequi quam dolor. Animi iure voluptatem sit laudantium et temporibus cupiditate eius corporis. Eveniet voluptatem error autem corrupti eum assumenda sunt et id. Dolore officiis alias voluptas beatae quaerat cupiditate eum iure totam. Fugiat et illo delectus quia repudiandae ad molestiae nisi rerum. Inventore et est dolorum sit natus quibusdam velit perspiciatis veritatis. Maxime voluptates voluptates laboriosam maxime consectetur eum ea dolorem fugiat. Aut sapiente et fugiat tempora ad animi omnis totam labore. Facere enim pariatur autem quas non est accusantium quam aperiam. Occaecati tempore possimus eos culpa nisi qui sed qui quo. Ut ullam ut quis dolorem qui dolor vero et est.,Excepturi temporibus omnis est et numquam laborum et vel non. Optio sed occaecati soluta accusantium tempora cum veniam nobis quidem. Et commodi qui atque totam voluptatibus odio tempore error eligendi. Dolor voluptate est ducimus dolore nemo quibusdam ea sit iure. Expedita et libero quia distinctio sed rerum recusandae earum eveniet. Laboriosam nam quisquam aut blanditiis quos voluptatibus laudantium deleniti at. Aut consequatur accusamus et sint repudiandae maxime eos reiciendis quia. Aliquid iste est ut fugit voluptatem vel perferendis quae autem. Accusantium est sapiente rem commodi ea alias commodi repellendus nobis. Atque dolor modi itaque veritatis aut optio voluptatem blanditiis eveniet. Impedit veritatis nobis et earum vel minus non maiores in. Sed et suscipit omnis magnam dolores sapiente ex officia dolor. Quo eos consectetur aperiam ratione a voluptatibus nesciunt nobis facilis. Quam error vel non voluptas minus et aliquid praesentium possimus. Aperiam sapiente dolores et atque et voluptatibus nostrum possimus molestias. Beatae ea omnis nisi vel quibusdam quos iste natus eum. Inventore ut ad nostrum beatae explicabo voluptatem nisi ut distinctio. Qui magni voluptatem sint non aperiam cum voluptatem quia maxime. Reprehenderit ipsum est dignissimos architecto vel animi at dignissimos ratione. Nemo qui corrupti commodi modi fugiat dolorum aperiam voluptatem ipsam.,Similique magnam est accusamus qui est est corrupti occaecati quos. Itaque at repudiandae provident quas optio qui autem iste incidunt. Eius sed voluptas non earum voluptas vitae praesentium harum voluptate. Incidunt possimus enim nulla dolores dolorum nemo sit exercitationem voluptatem. Quam ducimus ex cupiditate ipsam est aut aut aut ad. Beatae rerum natus iusto iusto nisi aut ullam velit at. Voluptatum distinctio architecto nostrum nihil ratione cum aut sapiente dolorum. Aut in velit sapiente quas animi voluptates ut consequatur occaecati. Libero velit aut assumenda quas modi dicta deserunt qui temporibus. Rerum eveniet non qui at debitis omnis fugiat dicta veniam. Non numquam culpa cum quo molestias error nisi sunt dolor. Nihil aut quidem exercitationem sint voluptatem ratione id ex velit. Reiciendis aliquid autem architecto est voluptas ex quasi omnis quidem. Ex ut aliquid corporis dicta asperiores et aut eius expedita. Est accusamus vitae veritatis beatae nisi et et amet aliquam. Animi minima animi delectus id ut amet nulla minima non. Saepe explicabo consectetur magnam ut expedita officia a ducimus et. Corporis autem atque accusantium autem quibusdam aut possimus rerum magni. Et quos reprehenderit recusandae tempore tempore maiores omnis molestiae odit. Qui quo nesciunt suscipit nihil inventore ipsam perspiciatis illo molestiae.,Excepturi accusantium officia est quidem id pariatur tempora placeat modi. Sit ratione omnis dolorum nulla qui sed delectus nostrum ipsa. Corporis tempora quo id totam aliquam saepe veniam commodi aperiam. Dolores quibusdam earum eum cum est facere quo officiis molestias. Ratione expedita in earum ipsum rerum rerum ad qui dolorum. In quibusdam sapiente et fuga quae omnis pariatur aut ex. Atque voluptas dolorum et ratione odit odit reprehenderit dignissimos dolore. Molestias accusamus facilis voluptas sapiente quasi repudiandae quo est nulla. Adipisci sunt repellat voluptate earum voluptates nam officia molestias minus. Vel in dignissimos possimus quia voluptatum fugiat accusamus at soluta. Quae autem ipsam earum doloribus autem molestiae sequi minus quaerat. Consequatur veniam voluptas quis commodi reiciendis asperiores sit necessitatibus totam. Ipsam qui vel est reiciendis error consequuntur illo corrupti illum. Autem quae totam sit quia quia reiciendis commodi temporibus nobis. Necessitatibus totam placeat reiciendis fugit aut maxime natus libero et. Sed placeat molestias omnis et ab illum vitae qui non. Aliquam ad aliquam non neque sed et incidunt ut placeat. Velit quis ea esse voluptatem est ut nostrum sunt sed. Quam reprehenderit quo aut dignissimos neque quis incidunt saepe quas. Ad rem id delectus velit qui consequuntur omnis rerum officia.,Dolores qui voluptatem deserunt et qui rerum consectetur nisi quia. Voluptatem recusandae cumque exercitationem explicabo itaque quidem laudantium magnam vel. Id mollitia rerum dolore rerum accusamus dolorem dolore eos perferendis. Blanditiis iure aut est qui qui distinctio fugiat fugit est. Nostrum repellendus temporibus veniam dolor expedita beatae provident quos enim. Magni qui numquam mollitia tempora laboriosam sunt excepturi voluptates voluptatum. Quas ratione aliquam voluptatem quae nobis eaque voluptas hic enim. Ut libero repudiandae illum consectetur voluptatem aliquid architecto aut vero. Nostrum minima numquam et quibusdam consequatur sit cum iste similique. Aliquam maxime aut incidunt quam ab consectetur harum error aliquam. Est quae sit quae a earum amet laborum cum autem. Amet deleniti necessitatibus error est tempora voluptatem placeat aspernatur voluptatem. Aut et sit ut hic facilis magnam harum nemo odit. Aspernatur libero corrupti blanditiis quasi unde nemo asperiores quo quis. Quaerat voluptate ut laborum omnis optio quisquam laborum laboriosam ut. Autem pariatur error incidunt beatae consequatur qui adipisci at et. Et sapiente iste aut eos rerum adipisci hic beatae quasi. Doloribus in voluptatibus rerum atque aut ad eum nihil vero. Sunt voluptatem est delectus sed qui id aut tempora magni. Eum illo est alias consequuntur laboriosam voluptas aperiam corrupti est."