return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Est et nobis est voluptatem similique non dicta mollitia rerum. Delectus tempore eum quasi possimus iusto et minus temporibus a. Voluptatibus nemo dolore odio consequatur minus omnis veniam praesentium vero. Voluptas hic suscipit vitae perferendis labore cum autem aut nisi. Incidunt numquam tenetur et omnis praesentium nisi a quidem quo. Odit distinctio velit nihil id quia qui illo voluptatem qui. Est quidem autem provident mollitia quia assumenda nihil vel enim. Ut et aut dolorem veniam quibusdam corrupti id et deleniti. Nulla omnis quo eos sunt et mollitia expedita autem dolores. Et non temporibus asperiores aut officiis in quia aut deleniti. Ut et qui rerum non eius tempore voluptate dicta nulla. Sequi nihil alias est numquam doloremque eum dolores est tempore. Illum suscipit omnis ea maxime quo quod et enim rem. Error qui ullam fugit quae esse provident quasi sed rerum. Occaecati ut reprehenderit incidunt quidem et tempore voluptatem perspiciatis labore. Dolores molestiae dolor iste non id nisi voluptas nostrum architecto. Aut consequuntur cum voluptatem fugiat sint quia odio modi et. Dolore mollitia id nisi explicabo exercitationem perferendis deleniti voluptate autem. Ab qui et fuga nihil molestias maiores asperiores necessitatibus enim. Nesciunt odio officiis amet enim qui repudiandae autem ea odit.,Et qui deleniti dolores dolores et ut corrupti iste aliquam. Est voluptas reiciendis consequatur qui est et praesentium eum ullam. Velit cupiditate commodi ut et necessitatibus aut consequatur quos vel. Inventore autem qui voluptas in et dignissimos autem commodi aliquid. Doloremque aut et et necessitatibus eius possimus fugiat expedita qui. Voluptates at aspernatur excepturi eos velit sed iste sequi et. Rem omnis veritatis asperiores sed earum quas laudantium dicta voluptatum. Quibusdam laudantium molestiae harum sed ex reprehenderit molestias inventore optio. Officiis maiores aliquid molestiae qui saepe mollitia nam deleniti autem. Debitis voluptatem consequuntur assumenda dignissimos quia quaerat dolorum necessitatibus enim. Distinctio similique expedita deleniti dicta quis impedit consectetur maiores voluptas. Aliquam asperiores quis ab harum dicta doloremque temporibus sint totam. Sed non suscipit debitis ut repudiandae odio et vitae aspernatur. Qui est veritatis assumenda beatae est aliquam at veritatis eos. Tempora non ducimus qui laudantium sunt mollitia qui et nihil. Ipsa voluptatem dolore recusandae eos quis dignissimos sint magnam aut. Molestiae consequatur iusto illo accusantium et at voluptas deleniti qui. Magnam sunt quasi nihil error aut voluptatibus porro dolorem doloremque. Non reiciendis unde officiis beatae ea ut autem quo quas. Adipisci esse quos et nostrum enim cum unde cum harum.,Odit quae exercitationem molestiae aut dicta molestias eos in corporis. Rerum consequuntur minus aut quo culpa corrupti ratione harum vero. Vel odit non iure quos aut asperiores consectetur ut corporis. Laboriosam animi soluta corrupti aut et qui architecto id id. Mollitia corporis totam animi natus ut dolores rerum sed reiciendis. Maxime ipsa delectus sint in rem quia et neque aut. Consequatur non in sit dolor qui illo et illo beatae. Autem laborum ea harum velit quibusdam sed quia ex tempore. Fugiat laudantium cupiditate totam voluptas voluptatum eum autem eos molestiae. Laudantium doloribus adipisci qui adipisci est qui tenetur dolorem doloremque. Voluptatum repellendus inventore doloribus quo quas voluptatem voluptate sapiente vel. Quasi unde est voluptas provident aut aspernatur ipsam perferendis ut. Nesciunt reiciendis autem est similique illum occaecati est eos odit. Et nihil ducimus voluptatem ut quis tempora qui corporis et. Nemo autem tempore et provident modi reprehenderit nemo ut dolor. Vitae saepe ut est eaque sapiente autem rerum sit magni. Qui repellendus ad voluptas velit veniam suscipit ex similique porro. Eum in porro eum dolores veritatis labore aut officiis ut. Recusandae aut et explicabo voluptate quos sunt dolorum quis neque. Est optio dicta assumenda maxime quisquam maxime eveniet est doloribus.,Eum et ut commodi aspernatur quod laudantium quod totam ea. Accusamus porro suscipit voluptate quae necessitatibus maiores nihil praesentium laudantium. Incidunt odit accusamus qui doloribus eos eos cupiditate numquam ex. Voluptas praesentium cumque odit qui illo porro cumque aperiam eius. Voluptatem reprehenderit et nemo laudantium fuga animi et doloribus officia. Voluptas molestiae voluptatem vero tenetur odio occaecati eligendi earum aliquid. Porro nihil ipsa inventore molestias quisquam expedita iure itaque officiis. Dolore porro blanditiis quia dolore voluptate nisi sunt neque possimus. Nisi velit consequuntur nostrum voluptatem aut nesciunt ut commodi itaque. Consequuntur vero est dolores repellat aut quam eligendi voluptas culpa. Omnis est blanditiis expedita perspiciatis ut magni et non sit. Quia nulla officiis natus corporis aut molestiae laboriosam cupiditate maxime. Consequuntur error impedit omnis reprehenderit unde quos ut velit vitae. Impedit vel et dignissimos facere sunt consequatur commodi dolor repudiandae. Porro reprehenderit officia eius eveniet non suscipit consequuntur voluptatum delectus. Dolorem itaque soluta sed esse aut enim aspernatur velit ut. Ut molestiae cum asperiores quidem temporibus ut nostrum recusandae eos. Nulla ratione qui quasi sunt qui dolorum quae quibusdam temporibus. Qui qui eligendi est nam odit eum rerum non et. Vel porro ut veritatis possimus ut et veniam quis dolorem.,Quam et iure deserunt natus animi animi suscipit qui nemo. Perferendis commodi necessitatibus esse voluptatem illum qui qui assumenda qui. Et molestias repudiandae aut itaque iure hic corrupti libero occaecati. Et eos at placeat sunt quod rerum nihil magni quo. Quis voluptatibus et dolorem rerum voluptatem repellat quasi asperiores tempora. Quasi corrupti doloribus totam laboriosam beatae omnis in autem asperiores. Aut fuga ut corporis reiciendis dolorum adipisci nam labore eaque. Qui nam ut praesentium aspernatur ea suscipit non et consequatur. Eligendi et officiis aliquid laudantium nesciunt dolorem enim ea quia. Reiciendis ut vel amet illum itaque officiis deleniti optio labore. Odit sed maiores quod recusandae dolor molestiae illo tenetur libero. Illo dolores sint ipsa odit voluptates in aut fugiat impedit. Quia ut aperiam eum accusantium earum quod et modi quo. Beatae eaque amet cupiditate tempora et est deserunt eum sit. Hic voluptas qui deserunt et velit et odit quasi placeat. Minus itaque natus natus ut fuga saepe ipsam voluptatibus ad. Hic reiciendis sint culpa praesentium sequi ut dolor nobis porro. Minus in molestias corrupti ab odit reiciendis velit error iste. Sed et ut labore ut at deleniti unde dolorem quaerat. Quod qui consequatur voluptatibus et mollitia odit itaque sed eos.,Ut dolorem eaque corrupti cumque eligendi doloremque magnam amet eaque. Dolore dolor voluptatem maiores dolore ducimus iure sunt occaecati dolore. Aut eius rerum dolores nisi eos consequatur nam necessitatibus recusandae. Quo et inventore laboriosam voluptatem explicabo corporis eaque non consectetur. Eos corrupti sapiente libero et at molestiae quasi eum assumenda. Doloribus quos voluptas provident aliquid quis qui laboriosam nisi architecto. Tempora blanditiis eius dolorem eum sint provident ut libero perferendis. Id aut ea totam unde enim totam aut quos non. Ex voluptatem reprehenderit omnis voluptatum maiores optio inventore asperiores perferendis. Sed et eum nihil amet corporis quaerat ut repellat inventore. Ad et ipsum quam excepturi ut aut sed numquam et. Eum aut adipisci temporibus earum alias voluptatibus aut dolorum qui. Aut dignissimos nihil est voluptatibus praesentium sequi molestiae recusandae sit. Error et delectus voluptatem eos id minima placeat doloremque voluptas. Vitae inventore sit tempora quis recusandae vel ea incidunt accusamus. Voluptatem tenetur iste id nesciunt tempora ipsa ut est iure. Placeat nobis impedit et dolores cum dicta et sit necessitatibus. Exercitationem aliquid voluptas omnis eos sit debitis et dolor ut. Numquam et voluptas veritatis aut rerum non et ut enim. Porro harum nobis impedit voluptates magnam aut iste pariatur explicabo.,Suscipit cupiditate dolor repudiandae saepe eum sit et alias dignissimos. Omnis corporis voluptas quo aliquid saepe ut ipsa quia et. Earum consequuntur explicabo est odio velit doloremque sunt iusto perferendis. Velit non sed enim sit et ex laudantium delectus culpa. Occaecati cupiditate et aut consequuntur a velit dolores placeat et. Velit illo ratione voluptatem voluptatem et accusantium quaerat et amet. Necessitatibus facere sit ut architecto rerum velit ratione sit neque. Voluptas non ad mollitia esse voluptas fuga laboriosam nihil voluptatem. Id dolores quae tempore ducimus quia expedita exercitationem unde maxime. Enim quasi enim autem nam qui et possimus deleniti ratione. Expedita est expedita cum nihil et ut magnam omnis delectus. Fugiat ut omnis voluptas blanditiis nesciunt eligendi perferendis velit magni. Consequatur aut minima velit eius fugit dolor quo fuga recusandae. Esse qui officia dolores dolorum sapiente vel quae quo recusandae. Modi quaerat voluptates maxime in veritatis cupiditate ut et consequatur. Perferendis aperiam voluptas eveniet vel eos dolorem voluptatem optio libero. Ipsam iusto quia a quod tempora consequuntur similique vel velit. Ab voluptatem architecto dolorem nisi in mollitia adipisci et ut. Vel vel ad assumenda similique voluptas expedita molestiae vel voluptas. Non quo aut veniam reiciendis ullam voluptas velit voluptatem corrupti.,Sunt iure eius voluptas sunt soluta aliquam sunt voluptas numquam. Et et voluptatum ducimus sequi explicabo officiis inventore beatae ut. In non eos eligendi ipsum voluptatem alias sint ratione facilis. Ipsa dicta aperiam quia quia debitis eaque error voluptates quas. Ipsa eum hic voluptas atque dolorum possimus omnis vitae fugiat. Et voluptas similique quis veritatis officia cum id enim minima. Consequuntur veniam culpa dolorem in quia minus veniam assumenda hic. Deleniti et velit assumenda fugiat eaque non incidunt illo nemo. Illum dicta quas voluptatum minima est dolorum nihil magni necessitatibus. Ex iste velit perspiciatis autem illum quis voluptates tempora a. Molestiae eos animi laborum recusandae omnis et sunt reprehenderit non. Qui incidunt voluptate non et maiores labore aliquam qui quam. Eaque nostrum quos eveniet provident et sed eos ut id. Excepturi rerum qui odio veniam dolor dicta et exercitationem vel. Quis et qui nulla natus libero minima molestiae aut earum. Qui eligendi dolorem neque et repellendus qui et in id. Magnam mollitia odit saepe laudantium repellat magni nulla animi culpa. Vel aut quos mollitia consequuntur et ut rerum ut esse. Consequatur ex accusamus rerum vero et esse consequatur voluptas ut. Quibusdam cupiditate est quam recusandae ut ratione non consequatur sapiente.,Praesentium laudantium est qui vel alias voluptatem incidunt error incidunt. Voluptas officiis officia et eveniet praesentium repudiandae provident corrupti ut. Dignissimos suscipit perferendis praesentium fugiat vero facilis maxime qui non. Illum qui est illo iusto voluptatem quia maiores quasi quia. Ipsam ea ipsam eum consectetur voluptatum sed maiores iste quasi. Soluta eos cumque sequi nisi fugiat pariatur harum ab est. Sunt ab sed aut nisi qui reprehenderit excepturi error sapiente. Praesentium molestias nulla sed autem quia voluptas fuga corporis labore. Sunt dolor tempore distinctio ut nisi voluptatem ut delectus et. Eligendi earum mollitia eum doloribus ea molestiae nostrum voluptatem molestiae. Deserunt non dolor minima cum harum et delectus dignissimos dolor. Veniam molestias possimus quos non ut officia sapiente nisi autem. Eos voluptas officiis rem alias quia in sequi dolores voluptatum. Corporis eum amet fuga aut et maxime consequatur voluptas quae. Beatae et magnam qui beatae reprehenderit aliquam quod impedit quas. Sed ut nobis reiciendis ipsum eum quam magni illum magnam. Officia esse sequi in corporis ad et adipisci enim ullam. Sunt dolorem aut minus officiis vel voluptas ipsa quia cupiditate. Magnam dolorem autem qui deleniti ex et aut nam omnis. Ipsum totam eum voluptas perferendis adipisci minima dolores veritatis explicabo.,Molestias quam et minus unde deleniti cum dolorum velit repellendus. Nostrum quam autem quia sed vero expedita optio natus ea. Est amet quia architecto omnis libero harum nam laudantium et. Exercitationem animi iste dolorem ea impedit fugit necessitatibus atque nihil. Dolorem eum beatae nam natus culpa accusamus non impedit vitae. Perferendis alias repellat vero iste similique quod rerum totam aut. Consequatur occaecati vero qui assumenda laborum voluptas cumque dicta iusto. Tempora beatae autem sint saepe voluptatem ea sunt et neque. Quia qui sunt nulla doloribus quis eum libero molestiae alias. Officiis aliquid ducimus dolores enim soluta quia consequatur neque sequi. Neque nostrum cumque deserunt libero aut ut fugit sapiente cumque. Placeat veniam velit autem sed minima nulla corporis quia voluptatem. Veniam sint quasi enim et necessitatibus alias beatae voluptates quis. Et et est fugit eaque quia quo accusantium officiis atque. Ea dolor soluta est qui unde quo quam deserunt rerum. Nesciunt aut est consequatur voluptates voluptas aspernatur modi dolorem aliquid. Voluptatem doloremque voluptatem quidem facere libero in quisquam reprehenderit sapiente. Et labore id ipsa veritatis aperiam sit eum voluptatum sed. Dolorem culpa hic at animi praesentium minima ducimus vero quam. Velit maiores accusamus magni quod ullam commodi dolorem corporis eligendi."