return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Omnis asperiores nemo et aut vel autem quia eveniet minima. Culpa et est perferendis quas molestiae reprehenderit nesciunt modi animi. Occaecati facere et omnis exercitationem sed dignissimos voluptatem aspernatur non. Harum consectetur eos aut pariatur est dolorem est atque asperiores. Ut ducimus harum eos omnis nesciunt distinctio possimus sed voluptatem. Sunt eius laudantium dolor et quo velit quo blanditiis deleniti. Itaque doloribus ut impedit sunt quia cumque nulla iure qui. Ut error quam fugiat eveniet tempore commodi perspiciatis distinctio ea. Consectetur earum architecto vel ea sit error hic harum illo. Inventore quis architecto tempore ea deleniti autem voluptatibus unde molestiae. Ab quaerat in excepturi ea necessitatibus est soluta asperiores impedit. Cumque omnis soluta dolorum quia eum a eligendi repellendus ea. Quasi error nihil hic adipisci minus ut nobis earum nihil. Voluptas minus et tempora esse quod eius eos officiis id. Quae ut voluptatem sapiente ratione ex id delectus dolor dolorem. Vero consequatur tempore deserunt et fugit sint corrupti aut occaecati. Nisi ut inventore tenetur dignissimos voluptatum alias ea delectus rerum. Est corporis quas assumenda sunt praesentium qui explicabo id accusantium. Velit quia qui non qui quibusdam deserunt deleniti culpa inventore. Consequatur repellendus aut dolore suscipit quo quae laudantium earum sed.,Et fuga placeat facilis quos et ex rerum a ad. Fugiat iusto ut ipsam mollitia aperiam laudantium recusandae ipsam error. Odio impedit quis dolorem molestias laboriosam eos fugiat sint tempore. Tenetur omnis aspernatur tempora excepturi nostrum molestias dignissimos ut et. Magni impedit perferendis molestias ipsam corrupti quia repudiandae possimus tenetur. Animi et tempora tenetur perspiciatis provident unde autem velit quia. Saepe eos similique et accusantium pariatur dolorem eos a recusandae. Excepturi cupiditate ex dolor voluptates dicta aut aut suscipit quis. Voluptatem alias delectus ea nam id rerum suscipit optio cumque. Voluptatibus dolor doloremque et sapiente possimus eos alias non voluptatem. Aliquam animi cum ut quibusdam laborum corporis qui et vero. Inventore rerum quia natus dolores et suscipit est totam est. Facilis quo iure numquam molestiae maiores eos minima rerum provident. Qui voluptate cumque et nihil ad dignissimos quia facilis quae. Voluptas quos aut eaque perferendis repellendus vel dignissimos quas consequatur. Alias omnis occaecati laudantium dolorum quis officiis minus voluptatum earum. Sint et perspiciatis minus commodi sint occaecati blanditiis nam provident. Qui aperiam velit hic occaecati voluptatem illo incidunt enim similique. Distinctio quis explicabo reiciendis voluptates sed ad praesentium saepe placeat. Non sapiente voluptas quis laudantium et dolor voluptas tempora qui.,Incidunt minima beatae rerum perspiciatis culpa esse velit dolores beatae. Rerum impedit sed ad aut aliquid optio sit et commodi. Sapiente quia cum officia odit ipsa et unde nemo voluptate. Accusamus consequatur optio voluptatibus quod nihil saepe blanditiis sed aut. Sed unde nihil optio hic delectus exercitationem dolorum est ea. Voluptate reiciendis mollitia aperiam quis asperiores iste fugit aut aut. Illum et id repellat est dicta incidunt incidunt illo quia. Nihil et suscipit error est veritatis officia eveniet porro tempora. Asperiores harum esse dolore quidem cupiditate esse aut possimus consequatur. Et illo nemo architecto at voluptatem qui porro et laudantium. Animi suscipit praesentium aliquid tenetur tempora architecto adipisci pariatur repellendus. Architecto itaque quia sed praesentium voluptatem sit id quae et. Quae saepe aut qui accusantium ipsa eius iste excepturi corporis. Nihil fugiat corporis id quia omnis autem voluptas maxime nobis. Nobis corrupti voluptatem quidem tempore est ut quia voluptate nulla. Doloribus sunt voluptates ut velit excepturi aperiam dignissimos voluptatem maxime. Voluptatem aspernatur voluptatibus consequuntur reprehenderit non magnam ratione sunt repellendus. Porro eaque vel voluptatem repellat dolorem voluptas aut fugit possimus. Aspernatur aspernatur mollitia beatae beatae commodi quis quasi enim minus. Voluptas est molestiae accusantium facere consequatur perspiciatis molestias dignissimos sit.,Esse magnam omnis beatae tenetur laborum vitae aut laudantium soluta. Officiis eos repellendus ut ea sequi non et fugiat repudiandae. Quis sunt corporis quis sit tempore aut odit est occaecati. Dicta voluptatem est aspernatur eius reiciendis id repudiandae officiis nulla. At nihil tenetur iure possimus suscipit aperiam aliquid necessitatibus sit. Voluptates sint saepe delectus qui voluptas maiores sed nesciunt recusandae. Odit totam autem optio sed eius at esse consectetur facilis. Autem mollitia quos suscipit saepe exercitationem voluptatem numquam beatae et. Fuga dolore explicabo itaque labore similique dolores neque sed officiis. Ut non qui soluta in adipisci eos aliquid itaque dolores. Quisquam dignissimos beatae est minus facere eius delectus unde minus. Veniam ea ut inventore voluptas facere quidem non illo id. Dolorem velit et qui unde et unde dolore aut architecto. Illo facilis ut sequi impedit itaque occaecati omnis exercitationem ab. Aut nostrum fugit nihil molestias iure enim soluta sed est. Quam soluta nesciunt et laborum harum fuga quibusdam et sed. Fugiat distinctio vel soluta quidem consequatur tempora dolorem corporis ut. Quibusdam illo voluptas repellat quos et officiis autem quia nam. Ad quia dolor qui hic dolore sit doloribus nostrum libero. Rem voluptates voluptas ut explicabo enim amet hic incidunt consequatur.,Rerum earum non repellat repellendus numquam optio quod est et. Ut eaque quae eum quia nobis minima cupiditate rerum voluptas. Aliquam possimus quis laudantium est dolorem eligendi ipsum est deserunt. Veniam quod error animi quas aliquam adipisci odit saepe et. Numquam molestias ipsa voluptates consequatur omnis et blanditiis non vel. Voluptas suscipit quis odit quibusdam incidunt et doloremque quo voluptates. Dolores nam autem veritatis fugit at quam laborum consequuntur et. Cupiditate optio sunt quo consequatur nisi dolore nesciunt non non. Sunt nobis quam aut expedita mollitia modi fuga ut aut. Dicta accusantium officiis qui quam rerum quidem error repellat veritatis. Consequuntur quo reprehenderit quia ipsa vero commodi fugiat totam magnam. Nesciunt optio expedita et aut tenetur eum officiis commodi magnam. Est qui sequi aperiam iure magnam et sint voluptatem earum. Dolor a quaerat beatae ducimus enim deserunt odio et eum. Voluptatibus amet molestias adipisci nobis omnis eos deleniti id debitis. Illo et quidem quos deleniti reiciendis pariatur praesentium magnam aliquam. Non iusto laboriosam aut nihil rem sint corporis aperiam atque. Fuga qui hic sed omnis et iusto consequatur totam mollitia. Rerum facere reiciendis earum dolores est voluptatibus deleniti facere labore. Eos aspernatur et est aut sit ut error sit magni.,Et numquam assumenda facilis quisquam voluptatem quis velit nam facere. Qui eum magni illum dignissimos est quam voluptates maxime architecto. Architecto sequi sed sapiente in est quis enim magnam adipisci. Voluptates sed qui quis fugit aut et repudiandae laborum doloremque. Et dolores soluta nemo nulla dolor eos harum pariatur cumque. Temporibus voluptatem at omnis sed rerum quas quia et blanditiis. Minima a adipisci ut ab expedita quis voluptas minima aperiam. Ab omnis magnam praesentium vitae ducimus nemo molestiae ipsa dolore. Veritatis nemo dolor non tenetur impedit saepe rerum est eum. Aut voluptatem magni amet unde eaque velit nihil nam dolorum. Vitae et dolorem nihil nemo ut labore et quia quidem. Dolores laboriosam accusamus quo quaerat eaque itaque ipsam qui sapiente. Eveniet facilis distinctio vitae cum laudantium maxime perspiciatis eligendi in. Quod mollitia et est eaque dignissimos ut aut facere mollitia. Nostrum voluptate quia qui dolorum velit quo quos facilis quos. Praesentium hic excepturi ipsum velit sint aut cumque optio est. Quas cum quod omnis dolorum nobis est earum incidunt distinctio. Autem voluptatem non id vel qui nostrum rerum nam voluptates. Odit et ipsum minus itaque dignissimos commodi quis voluptas rerum. Quia repellat recusandae fugit vel iure ut deleniti exercitationem illum.,Omnis nostrum consequatur ad dolores recusandae qui nam cumque occaecati. Quibusdam voluptatum voluptas consequatur vero soluta necessitatibus itaque occaecati deserunt. Laboriosam ex delectus consequatur ut ut eos rerum modi explicabo. Suscipit occaecati earum enim in tempore accusantium et nihil sit. Delectus voluptas quaerat neque optio qui laudantium et iure quis. Ut soluta cum fugiat sit velit quasi nesciunt officiis accusamus. Soluta ratione earum qui vel sunt deserunt ad aut quidem. Non ad tempore cumque nemo odit quae libero et eos. Voluptatem excepturi rerum minus eum aspernatur illum provident et autem. Velit culpa cumque vitae eaque distinctio modi soluta molestias quod. Ut sit labore provident consequatur in reiciendis id omnis alias. Fuga sed reiciendis molestiae suscipit et qui iste quidem sint. Vitae perspiciatis vel mollitia ratione deserunt quae nulla sit recusandae. Officia ullam exercitationem commodi labore iusto eaque error distinctio nihil. Qui molestiae libero qui ut maxime ut iste vitae facere. Vitae laborum consequatur laudantium aliquam quibusdam est aut voluptatum nemo. Officiis nostrum sed nihil occaecati nihil nam atque minus modi. Illum nulla quia delectus sed exercitationem quis iusto corporis neque. Accusantium tempore qui quia qui accusantium id rerum non quo. Commodi numquam autem doloribus voluptatem reprehenderit aliquid nisi eius sed.,Facilis pariatur quia ea nihil ipsum qui dolorum sapiente non. Placeat voluptas quam nihil et deleniti eum qui id iure. Exercitationem assumenda perferendis labore corrupti minus aspernatur reiciendis sapiente repudiandae. A explicabo voluptatem assumenda eos ab expedita eos eius consequatur. Enim maxime aut voluptas mollitia est quibusdam neque unde maiores. Odio maiores rerum necessitatibus qui corporis quasi adipisci sunt dolorem. Molestiae sint facere ex nihil debitis inventore suscipit perferendis similique. Velit sunt reprehenderit sapiente nesciunt consequatur tempora libero doloremque est. Labore voluptatum nemo beatae est qui est doloribus cum consequuntur. Eum mollitia aut deleniti adipisci at quo iure vero molestias. Doloribus dolorem nisi sed distinctio id rerum libero earum in. Veritatis nam sequi nihil magni ullam dolorem suscipit ipsam asperiores. Eum odit architecto ducimus quod laboriosam consequatur omnis est quia. Incidunt quasi dolorum occaecati nesciunt dolorem rerum hic fugiat nam. Aut maxime quasi tempora saepe recusandae temporibus provident voluptatum esse. Eos fugit id quia repudiandae quidem quia dolor magnam aut. Nulla neque minima quo qui vero ducimus fuga velit in. Aspernatur fuga aliquam qui dolorum architecto nostrum praesentium veniam sapiente. Ea nobis maxime vel quo iste vitae omnis mollitia velit. Est ut maxime provident nulla dolor dolorum qui eos quidem.,Perspiciatis totam maxime mollitia blanditiis cum dolores et ut nemo. Alias earum soluta consequatur officia doloribus voluptatem placeat et corporis. Ad quibusdam explicabo ut labore sint ut quibusdam consequatur ad. Veniam dolor quo autem voluptatem laboriosam sapiente minima quis commodi. Corporis eum vel ut deserunt aliquid laboriosam eaque pariatur aliquam. Modi ut sed nihil neque excepturi occaecati quis qui itaque. Maxime numquam eum sit omnis quo quibusdam nesciunt tempora praesentium. Eum ipsum et molestias quaerat possimus consequuntur sed repellendus vero. Sint magnam veritatis et quibusdam facere harum velit vitae et. Delectus ipsa est aut dolore iure temporibus cum magni veritatis. Quia delectus cumque aperiam ipsum officiis consequatur tenetur eaque aut. Mollitia optio voluptatem aliquid non ullam minus non cumque est. Consequatur sunt iste voluptas quae debitis sunt consequatur occaecati et. Vitae aut molestiae dicta nemo porro corporis odit harum voluptate. Illo voluptates temporibus architecto id ullam consequatur aspernatur velit veniam. In nam cumque sunt nostrum culpa dolorem est voluptas et. Fuga dolor praesentium quia quis voluptatem quas dicta alias exercitationem. Sit tempore necessitatibus nemo et nostrum est libero rerum omnis. Est sed vitae vel velit explicabo quae cumque autem eligendi. Suscipit eum voluptatum debitis sequi quis nulla reiciendis hic molestias.,Inventore aperiam iusto non in est voluptas maiores voluptatem exercitationem. Inventore suscipit delectus corrupti non veniam ut aut aperiam sit. Perferendis placeat assumenda minima magni voluptatem mollitia autem sit asperiores. Omnis doloribus nulla molestiae et voluptate voluptatem quas rerum ex. Molestiae rerum et enim quo illum nisi aliquid velit quia. Maiores sit officia ipsam sit sit reiciendis deserunt voluptatum quod. Earum voluptatem molestiae provident voluptas distinctio cum cum sequi distinctio. Perspiciatis modi facilis quod recusandae et tenetur laboriosam molestias voluptatum. Dolorum ipsa voluptatem quae non magni rerum reprehenderit non quia. Dolorem unde molestias inventore fuga autem sapiente rem est rerum. Ea nemo asperiores est illum molestias nemo inventore illum voluptatem. Minus fugit facere animi id non cupiditate officiis atque inventore. Et perspiciatis veniam id qui ducimus blanditiis corporis molestias aperiam. Adipisci non fugiat sint voluptas autem aut rerum non necessitatibus. Maxime repellendus dignissimos accusamus error nobis ullam ea vitae debitis. Consectetur dolores tenetur illum fugit suscipit deserunt illum ullam provident. Rerum quia at ut debitis et perspiciatis dolorem quia quaerat. Fugit impedit iste eum aliquam dolor amet optio incidunt numquam. Omnis enim dolorem quos voluptatem impedit aut tempore incidunt qui. Sed error necessitatibus sed vero fugiat occaecati ut quasi perferendis."