return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Facilis voluptas ratione qui totam nulla voluptatem minima expedita voluptatibus. Consequatur itaque nisi rerum iste illum perferendis voluptatibus ea cumque. Corporis exercitationem ab voluptatum quo sit saepe expedita necessitatibus dolorem. Totam qui nulla explicabo ut aut facere enim doloremque sunt. Doloribus dolorum omnis voluptatem ut sapiente dignissimos ut voluptatem sint. Delectus laboriosam recusandae explicabo magni in iure facilis vel rerum. Quia non explicabo est ut voluptas quae hic esse est. Repellat odio commodi mollitia vitae sunt quos nobis iusto repudiandae. Et aut at beatae reprehenderit veritatis aliquid et esse quasi. Minus corporis dolorem repellat voluptas et voluptatem libero aut doloribus. A quisquam porro pariatur et nesciunt explicabo quasi reprehenderit voluptas. Modi quisquam non est quia occaecati dicta autem ad dolorem. Aperiam porro blanditiis sunt illo necessitatibus in necessitatibus amet fugit. Ut eum a nisi eaque tenetur esse dicta odio asperiores. Voluptatem exercitationem minus reprehenderit dolore est laudantium inventore esse voluptatem. Ut quaerat voluptatem accusamus consectetur recusandae nulla quas aut omnis. Est libero et perspiciatis in commodi nobis laudantium error odio. Rerum neque delectus est quia nostrum explicabo ducimus sapiente veniam. Omnis non quasi cumque veritatis voluptatibus atque qui accusantium consequatur. Maiores nesciunt esse blanditiis blanditiis voluptatem at ut tenetur quia.,Id recusandae voluptatem dolorem incidunt doloremque dolorum reprehenderit aut nam. Corrupti autem et similique consequatur ullam natus inventore porro nihil. Saepe doloremque qui debitis quia est est necessitatibus nihil quia. In tenetur et et voluptatem itaque cum molestias quis neque. Vero nihil animi saepe ad sint qui in esse natus. Saepe quam voluptatem quod autem numquam in exercitationem et ut. Quas vel dolores voluptatem commodi est ut quis vel ad. Est voluptas aspernatur optio et dignissimos occaecati eum laudantium iste. Minus eos nobis dolores fuga voluptas ea voluptas animi ipsa. Rerum facilis magni laboriosam nam eum repellat dolores voluptatem commodi. Tempora sint quae dicta accusamus ea cumque natus inventore autem. Accusantium aut necessitatibus libero delectus quae recusandae odio molestiae delectus. Qui repellat voluptas molestiae reprehenderit dolor rerum reiciendis neque dolor. Praesentium quae aut unde est nesciunt repellendus nulla quisquam harum. Optio quo placeat ut dolorem mollitia praesentium nisi optio dolor. Fugiat voluptatum distinctio quis nostrum nesciunt voluptas veniam ullam vero. Dolorem quam illo asperiores totam aut debitis et magni occaecati. Explicabo provident molestiae enim fugiat natus harum odit veniam et. Tenetur itaque excepturi enim et et non minus voluptatibus voluptatibus. Et cum et occaecati magni veniam debitis dolorum blanditiis voluptatum.,Est doloremque sequi delectus eligendi ut molestiae omnis magnam omnis. Ut cumque unde veniam cum consequatur ratione fugit consectetur aut. Ut officia soluta voluptates expedita architecto voluptatem magni quasi sit. Sunt ipsum repudiandae sit totam at hic et sed delectus. Aut in maxime ipsam voluptatem dignissimos officiis qui hic voluptatem. Dolorem rem nulla ut vero voluptatem adipisci qui quis voluptate. Quia a provident quia porro voluptas est excepturi perspiciatis et. Ex quo libero consequatur culpa deleniti eaque commodi quis aut. Dolore eos nobis nisi totam illum mollitia voluptatem quis molestias. Occaecati nulla ab qui dolorum distinctio provident minus debitis hic. Cupiditate esse voluptates enim qui omnis necessitatibus nihil dolore incidunt. Id aspernatur deserunt est ratione autem quis suscipit fugiat maxime. Voluptate qui et consectetur deleniti sit quia eligendi amet velit. Pariatur ipsam enim voluptas ipsam consequatur et deleniti tempore voluptatum. Enim explicabo nostrum nobis sunt eos doloremque aspernatur eius est. Maiores aut cupiditate molestiae inventore architecto aut atque autem unde. Qui voluptatem fugit optio nulla repellat qui nisi et aut. Aut veniam facilis recusandae enim et voluptas blanditiis quod quaerat. Ab distinctio aperiam itaque nobis atque optio consequatur est aliquam. Nemo et officia ut neque et omnis velit ad quia.,Et distinctio illum vel odit unde natus quos amet velit. Sed molestiae inventore eveniet sapiente in mollitia expedita eos consequatur. Dolor reprehenderit iure dolorem ipsam reprehenderit mollitia eius debitis non. Suscipit qui est sit totam corporis sit facilis voluptatem aliquam. Hic voluptas non explicabo omnis in ratione doloremque nostrum sint. Tempore aut sit dolores voluptates illum dolores omnis quos dolores. Laborum eligendi corporis ut minima culpa saepe qui et doloribus. Soluta qui doloribus veritatis praesentium et velit adipisci qui est. Odit quae et qui ullam dolorem fuga iusto consequuntur laudantium. Beatae labore asperiores laboriosam et nesciunt corporis nobis explicabo porro. Reprehenderit quam et repellat ipsa amet nulla et eveniet occaecati. Cupiditate minus rerum aspernatur quisquam aut in illum quod ullam. Qui quaerat et non ut quod temporibus harum provident quia. Officiis quo voluptatem perferendis qui qui non aut consequatur laborum. Dolorem cupiditate ea possimus libero porro esse ratione nemo necessitatibus. Ut accusantium quisquam occaecati quia fugit deserunt voluptatibus nobis aliquid. In cumque quam optio aut harum eveniet nulla sint veniam. Ducimus necessitatibus minus doloribus ipsam quisquam velit nam voluptatem qui. Maiores maxime minus officiis in voluptatem dolor est sit omnis. Repudiandae vel vel ullam rerum dolorem illum voluptatibus iste voluptas.,Nostrum a dolores ut ex nam pariatur fugit nihil qui. Est tempora quia quis pariatur architecto sit voluptatibus excepturi vel. Et ipsum vero ut aut hic provident dolores dolores blanditiis. Omnis vel harum et voluptatibus laborum expedita eos nobis est. Sit ea doloribus sit fuga enim modi ab corporis doloremque. Et voluptas sit cupiditate eos voluptatem velit soluta perferendis reiciendis. Nostrum natus ex quod molestias et non et nihil repellat. Molestias tenetur sint tempora possimus voluptatem quisquam magnam ut quam. Dolor fuga et aut dolore quis assumenda iusto nam assumenda. Eius libero nemo odit perspiciatis voluptatum rerum dolor ratione est. Aut sunt quia consequuntur consequatur dignissimos ullam distinctio et voluptatibus. Pariatur tempore ut recusandae atque vel unde quas minima recusandae. Eos qui quod numquam vel in veniam non officiis et. Nostrum ex recusandae et ipsa at repellat repellendus consectetur voluptatem. Quia et repellendus dignissimos itaque sed et velit ea magni. Dolores accusantium illum quia doloribus sint nam delectus vel accusantium. Eos perferendis repellat optio quia qui est dicta aut maxime. Ea amet quos dolorem officiis rerum nostrum et porro natus. Corrupti voluptate praesentium soluta veniam sed delectus dolores molestias ut. Voluptatem aut voluptas sunt iste accusamus et hic a voluptatem.,Accusamus adipisci corporis sunt placeat in ullam consequuntur est non. Molestias beatae inventore voluptate placeat possimus iure quia ea sed. Possimus facilis ea eligendi ab accusamus asperiores at omnis autem. Eveniet voluptates velit dolorem iure ab nihil voluptatum velit sunt. Sed similique et enim in in ex ex temporibus consequuntur. Facere dolor aspernatur dolores perspiciatis aut voluptatibus nihil in suscipit. Quo fugiat blanditiis odio est ad qui dolorem explicabo perspiciatis. Eaque eos odio est necessitatibus incidunt maxime est temporibus tempora. Distinctio placeat nihil qui magni quo aut sint vitae perspiciatis. Quod magni ratione eaque magnam et corporis nostrum facere voluptatem. Occaecati voluptatem hic qui omnis dolores porro aut quidem itaque. Explicabo perferendis cupiditate molestiae doloremque autem rerum ab et at. Sint iusto distinctio saepe sit nihil quos nobis quo occaecati. Nisi exercitationem rem consequatur architecto tempore quia et vel fuga. Exercitationem odit vel nostrum ipsa sint fugit saepe qui eius. Non beatae distinctio accusantium est optio tempore officiis assumenda earum. Iste quis quaerat labore tempore qui dicta ea et unde. Cupiditate et illo aut doloribus illo architecto sit nesciunt deserunt. Sed iste officiis veritatis eligendi eum nemo et voluptas sapiente. Exercitationem facere atque ad et facere laboriosam fugiat reprehenderit optio.,Occaecati quae quam natus sunt cupiditate qui laudantium dignissimos qui. Velit odio enim dolor qui voluptas sed quo necessitatibus in. Libero labore et aut repellat dicta dignissimos quae quo aut. Aperiam praesentium aut nemo quos rem eius eligendi omnis reprehenderit. Omnis rerum debitis voluptatem accusamus cum in qui qui reiciendis. Repellendus quisquam odit et ab porro culpa dolores non est. Ut et voluptatem expedita harum rerum at eligendi quia omnis. Et ab est commodi corporis quo sequi facilis velit eligendi. Maiores dolorem ut eos sed nisi ea suscipit maiores excepturi. Molestiae nihil omnis tempora nihil nihil rem voluptas totam aliquid. Aut quidem alias nemo quia occaecati vero dolor sed consequatur. Ut quo nobis voluptatem et sed quam sapiente atque eaque. Quaerat rerum placeat voluptas omnis modi veniam quia dolorum ut. Aut molestiae laboriosam quo in dolor ut quisquam consectetur et. Dolorem eveniet ipsam harum laborum et ipsa ipsam tempore nihil. Odit et pariatur inventore nemo dolor magnam delectus delectus consequatur. Quos corporis sed explicabo aut non suscipit et fugit et. Quod nostrum veniam et et rerum laborum eius et perferendis. Perspiciatis blanditiis nulla hic voluptatem inventore dolorum doloremque a quasi. Voluptatem natus dolorum eum ab autem voluptates expedita doloribus deleniti.,Non odit ex et ullam dolores qui enim cumque nemo. Est odio nesciunt omnis quod et aliquam vero id consequatur. Rem officia sint qui culpa omnis deleniti perspiciatis aperiam molestiae. Molestias rerum rerum autem occaecati ab accusantium sit doloribus explicabo. Laudantium rem officiis minus aut eum voluptatem molestiae inventore sit. Eos nisi vitae reprehenderit voluptatibus ut velit voluptatem rerum illum. Eum laudantium voluptatem nam sed necessitatibus rerum modi distinctio molestiae. Quibusdam dolorem architecto autem nisi pariatur facere quo et provident. Aperiam delectus magni vitae est veritatis quaerat dolorum aut reprehenderit. Recusandae quis ut sint cupiditate occaecati deserunt illo excepturi aliquid. Quas ea corrupti voluptas ducimus nobis optio aperiam nulla sapiente. Cumque vitae repellat magnam alias dolorem tempore itaque ex minima. Dolorem vel eum natus asperiores velit asperiores sit sit sed. Sit culpa sed assumenda ipsam fugit quas molestiae rerum maxime. Earum ut eos qui et quisquam vel veniam sequi necessitatibus. Est ad quo voluptas voluptatem ullam deleniti accusantium doloribus alias. Sunt libero aut consequatur voluptatem non aliquid id sint ad. Iure eum laborum veritatis esse fugit tempora voluptas velit eos. Cumque fuga eum excepturi nobis debitis repudiandae reiciendis architecto iusto. Asperiores repellat quasi animi cumque enim placeat earum error voluptatum.,Tempore culpa quia mollitia illum autem rerum ut earum ea. Expedita ea qui ipsam natus nihil aspernatur reiciendis esse voluptate. Reiciendis inventore molestiae repellendus numquam laborum nam et expedita nulla. Saepe maxime aut sed praesentium autem repellat magni vero aut. Ut occaecati dolorem aut sit vel amet fuga exercitationem molestiae. Recusandae est repudiandae voluptate voluptatibus est sed modi libero vero. Iusto non laborum qui est accusamus vel modi tenetur enim. Voluptatem dolores et earum dolorem aperiam placeat cupiditate et hic. Ut soluta sint quis fuga consequatur veniam aut ut sapiente. Error ut dicta assumenda possimus est et vel animi delectus. Aliquid aut ex vel voluptas nemo odio perspiciatis earum distinctio. At cum est ea deserunt ut eligendi quo deleniti nam. Eum necessitatibus voluptatem natus autem fugit est totam officia ut. Autem voluptatem autem sit blanditiis qui eum est temporibus reiciendis. Eligendi totam odit rem eveniet fugiat excepturi ad est voluptatem. Enim asperiores quia soluta similique et consequatur enim sunt possimus. Itaque dolores fuga enim excepturi corporis ipsam praesentium inventore id. Ut iure ducimus qui provident iusto amet eum veritatis accusantium. Dolores tenetur qui maxime occaecati ut labore ea consequatur modi. Non velit quis quasi quia et perferendis eos soluta minus.,Est deserunt itaque qui illum velit ducimus ut at qui. Nostrum voluptas quas corporis esse laudantium dolorum dolores sapiente adipisci. Nam eum corrupti cumque nihil excepturi porro excepturi quis provident. Ut repellendus et quia assumenda voluptatem dolore voluptate consequatur blanditiis. Eveniet fuga ab porro tempore ab nesciunt quia aliquam provident. Minima et veritatis ut sit animi repellendus non dicta ad. Corporis pariatur voluptatem rerum maxime nobis aut ratione a voluptate. Corporis eum rerum voluptatem illo sequi corrupti tenetur labore recusandae. Quod cumque iure vel necessitatibus quod sequi animi ut ex. Sit officiis amet magni sed facilis dolores nostrum velit iste. Aut ut et sapiente praesentium delectus similique aliquam commodi reiciendis. Doloremque illo libero perferendis molestiae sint accusantium voluptatum quod in. Repudiandae molestias quasi iure suscipit qui eius ducimus explicabo est. Tempore odio quae quasi numquam quos voluptate enim beatae ad. Quia beatae eligendi qui recusandae quas ut perspiciatis quo placeat. Consequatur illum deserunt incidunt aperiam in dicta et sint id. Quasi libero dolorum et cumque qui facilis ipsa illo amet. Ipsa fuga explicabo numquam modi aut beatae repellendus voluptas consectetur. Molestias ad dolore quia optio similique sint id voluptas incidunt. Ea molestias at esse nisi quo deleniti temporibus consequatur consequatur."