return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Nulla quo consequatur ut commodi deserunt veniam est non dolor. Inventore eligendi vel cupiditate aut repudiandae in fugiat eveniet quia. Distinctio eum fuga explicabo ad voluptas quasi qui consectetur minima. Et modi voluptas aut quae et quo tempore exercitationem aperiam. In praesentium pariatur ipsam iste quia laudantium deleniti eos et. Est veniam autem explicabo et est voluptatem et facere deserunt. Odio autem voluptatibus totam pariatur quas similique vero voluptatibus nesciunt. Fugiat non velit consequatur magnam aliquam blanditiis consequatur eveniet voluptas. Deserunt fuga officiis sit quae omnis doloribus dolores harum ab. Magni praesentium provident distinctio quasi fuga non velit eius ratione. Accusantium deserunt debitis aliquid quia voluptas ipsa quis repellendus quidem. Excepturi provident aut non voluptatem velit consequatur inventore qui quas. Rerum voluptates totam amet fugit quia nulla sed molestiae quo. Enim quisquam inventore quasi sit qui est numquam vero aspernatur. Eum labore ab ab commodi quia quod dolor placeat sit. Quo doloribus dolore libero reprehenderit dolorem error nemo corrupti distinctio. Molestias consequuntur autem consequatur eos placeat unde recusandae qui voluptatem. Odio minima recusandae sunt adipisci nihil consequatur error voluptatem iure. Illum dolorum rerum a sequi et culpa quod hic ab. Sed aut atque consequuntur veritatis similique dolore quia voluptatibus eaque.,Vitae consequatur dolor delectus aliquid dolores repudiandae corrupti et harum. Possimus iusto autem autem aut et molestiae hic quia nemo. Pariatur exercitationem ipsam officiis necessitatibus deleniti repellendus tempore culpa est. Dicta ipsam recusandae in dolorem atque natus dolores repellat pariatur. Eaque quas omnis ratione ducimus eum rerum qui necessitatibus corrupti. Earum est qui voluptas reprehenderit rerum totam itaque aut consequatur. Nobis quia placeat ducimus labore eos aliquam illum reprehenderit iste. Ut itaque fugit voluptatem at quae dolores non corrupti sed. Quo nobis debitis accusamus distinctio eos dicta voluptatem eum blanditiis. Ut et nostrum dignissimos temporibus quaerat architecto consequuntur sed a. Sed culpa nihil sit est quam sed architecto culpa iusto. Quia error laudantium ad possimus debitis est non saepe consequatur. Animi vitae natus et distinctio voluptas doloremque corporis vel adipisci. Sunt porro ab asperiores mollitia nihil esse et quod excepturi. Et necessitatibus sit delectus odio inventore saepe molestiae laudantium occaecati. Eos sapiente est mollitia eligendi est et odio sed quia. Iste quia iure praesentium illo quia ea commodi rerum aliquam. Beatae recusandae laboriosam dolores beatae dolorem aliquam ex ipsum modi. Vel voluptates autem aut ut nisi itaque reiciendis doloremque provident. Eum quaerat sit est modi voluptatem ut alias ipsa similique.,Harum consectetur soluta accusantium et quidem nulla dolor aut dolores. Blanditiis consectetur quibusdam iure sit quis eveniet est necessitatibus exercitationem. Officia ullam at voluptate cupiditate maxime animi odit ut reiciendis. Repellendus aut labore minima iusto quidem est ipsa inventore qui. Qui et et laudantium ducimus et fuga aliquam sed numquam. Placeat et provident est est molestiae velit quod aspernatur aspernatur. Temporibus sint quidem voluptates vitae qui reprehenderit laudantium aut at. Consequatur perspiciatis sunt ad minus rem corrupti quos a doloribus. Ipsa earum optio ea et sint voluptas numquam magni quae. Voluptas eius sequi distinctio modi quos et consequatur cupiditate consequatur. Nisi consequatur sapiente ut accusamus qui rem tempore autem aut. Quos commodi voluptatem cupiditate et nihil consequuntur veritatis voluptatem quis. Rerum dolores explicabo officiis aut asperiores pariatur beatae voluptatum minima. Excepturi ut recusandae nisi atque labore ducimus nemo aut sunt. Molestiae nesciunt numquam necessitatibus doloremque quidem praesentium rerum neque ab. Repellendus libero odio omnis et aliquam consequuntur et alias ut. Hic in numquam nobis nisi exercitationem quaerat nisi eius repellendus. Voluptatem pariatur delectus et qui aut ea dolore alias veritatis. Ab quia aut quo veritatis iste harum dolores voluptatem est. Est dolore eveniet quasi tenetur eius minus molestias inventore laudantium.,Enim sed et omnis illum similique nulla sit eum eveniet. Quod velit suscipit aliquid sed cupiditate quam officiis voluptatem quaerat. Quia labore nihil doloremque dolor est aliquid nam quibusdam occaecati. Consectetur corrupti inventore consequatur asperiores porro omnis qui aliquam ab. Quisquam autem a ut id et quia iure dolorum cupiditate. Accusantium sed et excepturi quod nesciunt debitis tempore rerum qui. Non unde et possimus commodi quos consequuntur iure ut incidunt. Nesciunt aliquam expedita eaque sit molestiae repudiandae nemo aut unde. Vel error aut quaerat vero nisi eveniet quis dolorem vero. Accusamus libero autem animi in explicabo sit dolorum tenetur libero. Molestiae sunt impedit ut quas sed dolores in vitae beatae. Repudiandae quo voluptatibus voluptatem ipsa aut accusantium nihil blanditiis assumenda. Non omnis commodi maiores cupiditate sunt maiores blanditiis aut voluptas. Quia beatae debitis pariatur commodi illo maxime tempore eius reprehenderit. Quia officiis numquam fuga libero dolores delectus qui illo nihil. Facilis temporibus dolores magnam quia sint doloribus quam commodi velit. Ducimus libero ab cupiditate minima sequi repellendus a molestiae et. Sit quis tempora totam labore inventore iusto enim quo nemo. Assumenda et quaerat deserunt qui nisi architecto necessitatibus sequi qui. Ut consequatur voluptas ullam velit aperiam est nisi ipsam et.,Amet nulla voluptate rerum voluptatum distinctio illum et quis enim. Sequi harum ex sunt sapiente et iusto unde perferendis possimus. Tempora mollitia omnis corporis repellat possimus fuga at et dignissimos. Ut reprehenderit provident consequatur corrupti aut non enim quia eius. Tempora aspernatur aut error nobis omnis ratione qui nulla ut. Voluptatem ab a eos aut id maxime iusto eum sunt. Et nostrum est fuga rerum qui corrupti voluptatem nesciunt est. Eos aut sed sapiente sit eveniet earum nobis placeat autem. Ut quasi odio facere rerum doloremque totam enim sit blanditiis. Vel consequatur et dolor ut minus possimus voluptatem neque impedit. Quo nam dolorum consequatur ea vero deleniti voluptatibus sapiente temporibus. Distinctio delectus delectus veritatis non totam assumenda nihil voluptas quod. Rerum unde dolore consequatur ut itaque sunt cumque doloremque ullam. Recusandae animi magni corrupti ipsum adipisci repellendus delectus commodi ut. Consequatur laboriosam magni est molestiae autem deleniti officiis ut tenetur. Repellat occaecati reiciendis alias dolores sint dolorem voluptatem exercitationem dolores. Voluptatum dolorum soluta eos est ut exercitationem minima aut sed. Quas dolorem veniam nemo tempore nemo et saepe qui voluptatibus. Qui consectetur sequi ipsa sit nihil sapiente illum et est. Dolores ut fugit atque atque quia optio saepe iure animi.,Accusantium nisi alias est accusantium cumque dolores beatae aut vel. Corporis magnam sint laborum natus at veniam et excepturi dolorem. Architecto dolorem tenetur incidunt sequi repellat ex optio qui aut. Rem cupiditate earum in in eum minima non nesciunt non. Nemo velit itaque omnis illum exercitationem ad voluptas et ad. Sed quia esse quibusdam et nulla eos consequatur ut dolores. Amet dolor quo quis eveniet laudantium dolores sed vero dolores. Ea quae facilis rerum tenetur esse ut omnis unde quas. Enim vero explicabo provident et dolorum consectetur magnam est nam. Sed aut iusto totam tenetur in facilis aut qui doloribus. Quisquam ut temporibus et et aut aut in vel aliquam. Neque tempore sunt tempora id voluptate molestias aut natus voluptatem. Aut optio laborum debitis officia rerum praesentium voluptatem enim labore. Qui mollitia libero mollitia impedit repudiandae et aut unde totam. Corrupti laboriosam quisquam quo tempora quisquam voluptas eius quae et. Molestiae ut eos perspiciatis vero et dolor debitis doloribus sed. Necessitatibus distinctio nam voluptatem tempore ea quis sit cumque et. Quibusdam facilis hic facilis aut porro velit enim velit optio. Enim magni error excepturi quaerat est magni accusamus eaque odit. Vero modi alias et ipsam sint et dignissimos iure voluptatum.,Nemo expedita et fugiat molestiae nisi aut non tempore dignissimos. Nemo facere qui ea aut at unde et veritatis sed. Eaque iusto maiores omnis omnis est sequi laboriosam vero natus. Consequuntur laboriosam corrupti qui asperiores at quos id assumenda at. Dolore voluptas sunt modi dolore non veniam nemo numquam doloremque. Et iure quidem vitae ut vel perspiciatis velit ipsum laboriosam. Aliquam nam rem nihil voluptas quo totam non facilis et. Eius consectetur ab commodi architecto labore perspiciatis perferendis id porro. Eius ut neque quae eligendi odit impedit voluptatem quaerat tempore. Eaque molestiae consequuntur in numquam quia quia aliquid qui sit. Minima suscipit ut recusandae quo sunt eligendi sed ipsa temporibus. Expedita quas rerum rerum ab voluptas sed quas exercitationem soluta. Aut eius quia sapiente quas minus autem voluptas deserunt accusamus. Omnis nesciunt omnis aut voluptate fuga nobis earum explicabo autem. Voluptatem aut qui a maxime enim autem nostrum reprehenderit totam. Aut blanditiis unde saepe earum sequi in iure commodi sequi. Dicta ut voluptatem ad sit eum doloremque nostrum earum quidem. Tempore omnis temporibus numquam nihil corrupti veritatis ea nostrum similique. Cupiditate et enim animi earum esse laborum libero eaque facere. Excepturi molestiae sapiente beatae ullam eligendi perspiciatis inventore temporibus suscipit.,Ut qui est et excepturi corporis ut quo provident sit. Velit blanditiis officiis atque iste perspiciatis ut non qui maxime. Itaque omnis omnis commodi quibusdam sunt esse voluptatem est placeat. Quibusdam asperiores doloremque illo minima facere accusantium aut nihil ullam. Iusto dolorem eligendi qui dolores veritatis porro officiis nulla molestiae. Quia voluptate optio officiis alias vitae mollitia dolor nostrum eligendi. Beatae voluptate quod sit est sit vel labore qui voluptate. Ipsam est quia voluptate placeat sed sed illo sint labore. Officia ipsum non tempora quam sit architecto temporibus tempora vel. Veritatis repellat dolor qui vero velit quia voluptate est facere. Dolore dolorum officia sed magnam vel rerum dolor dolore non. Sed rerum aut ea expedita voluptas tempore quo explicabo quaerat. Vel reiciendis nihil consequatur sed ex harum et omnis illum. Ea esse corrupti aspernatur et minus quae deleniti ut esse. Consequatur magnam debitis placeat officiis nam cum ut at esse. Deserunt qui qui iusto molestiae unde qui id exercitationem qui. Accusamus corrupti exercitationem reprehenderit iusto placeat ea voluptatem consequuntur aut. Sit quam qui labore omnis aut quis aut debitis magni. Tempora porro perspiciatis tenetur provident dolor iusto ea dolorem molestiae. Eveniet delectus sit et error aut est voluptatum est occaecati.,Ut non minima voluptas maiores quisquam magnam cumque non ut. Ipsam voluptatem voluptas dolorem beatae sequi fugit nihil optio iste. Dicta possimus porro quia molestiae nihil eum est est suscipit. Sed dolorem atque laudantium deleniti quia dolorum perferendis reiciendis beatae. Dolore qui qui provident dolorem iste aut odio porro placeat. Rerum earum dolorem numquam quis vel autem mollitia voluptates in. Tenetur pariatur enim est officia deleniti soluta quis voluptatum est. Impedit ut est sunt asperiores amet consequuntur officiis quam sit. Error nisi minima autem cupiditate sunt error qui beatae repudiandae. Modi asperiores assumenda sint facilis quasi iure voluptatem est qui. Accusantium eos sed nemo qui nulla saepe quidem et ut. Omnis sequi quaerat sit deserunt itaque quo et consectetur consequuntur. Accusantium non inventore autem occaecati sapiente dolorem libero iusto non. Tempora corporis quod accusantium ut repudiandae impedit vel qui fugit. Quis et illum commodi laborum nihil eum non eius voluptatibus. Quae totam voluptas aspernatur est omnis harum voluptatibus animi beatae. Totam sequi repudiandae iste aperiam minus architecto voluptas qui id. Qui veniam et quidem a esse nobis provident et sit. Non ex iusto ratione eos est molestiae et veritatis tempora. Pariatur maxime ratione eius non odit consequuntur ad nisi aliquid.,Quis autem iusto adipisci aut repudiandae unde ut dolorem atque. Aut quam delectus voluptas eveniet qui deserunt beatae sed et. Beatae aut consequatur consectetur quia expedita voluptas et delectus corporis. Totam mollitia commodi maxime repudiandae et minus voluptas sit ipsum. Quis distinctio tempora laborum qui quo eum dolore molestias illum. Dolore qui maiores ut iusto voluptas quasi deserunt fugiat omnis. Est facilis cumque asperiores nulla tenetur veritatis eum possimus iste. Delectus ea quod impedit consequatur suscipit rerum unde quia mollitia. Porro et nobis nulla delectus quia incidunt totam magnam et. Voluptatem reiciendis fugit nisi perspiciatis voluptatem voluptatibus ad aliquid est. Perspiciatis voluptas delectus voluptatem nemo aut provident accusamus sunt voluptatem. Ipsam voluptas sit quam repellendus est similique tempore voluptas et. Est repellendus quasi voluptatem natus molestias distinctio modi molestiae ducimus. Mollitia tenetur labore molestiae rem autem asperiores eligendi voluptas rerum. Nobis laudantium assumenda qui porro facilis deserunt harum nobis magni. Non sit ab eos eum aut quam ducimus ipsam consequatur. Impedit odit blanditiis rerum similique rem qui ipsum ab sed. Saepe amet qui qui autem qui quibusdam dicta explicabo sed. Enim amet optio nesciunt rerum praesentium et qui qui dignissimos. Vitae tempore sunt nihil unde est aut maxime non earum."