return a paragraph

return a paragraph

get_paragraph(count, sentence_count, word_count, separator)

Examples

get_paragraph(10,20,10,",")
#> [1] "Praesentium sed ut inventore magnam labore quia rem possimus aut. Voluptatem non itaque enim laudantium velit vero earum totam non. Sunt sed sed sunt nisi eligendi impedit harum sapiente maiores. Eum recusandae expedita vel sit voluptatem placeat corporis quod consequatur. Cumque quasi voluptates voluptate earum dolorem consequatur vel doloribus temporibus. Nam minus sunt error aut explicabo rerum ut exercitationem laudantium. Id alias sit facere velit et possimus soluta at vero. Et suscipit voluptatibus reprehenderit minima velit minima qui architecto doloribus. Sint quo ut dolor praesentium enim deleniti aliquid natus molestiae. Nisi aliquid debitis odit doloribus ipsum deserunt vel repellendus labore. Recusandae soluta voluptate quos in illum nobis labore doloremque unde. Laboriosam eligendi nostrum atque occaecati autem unde nam aut aliquid. Excepturi fugit consequatur maiores eaque molestiae sed aut ea id. Veniam quibusdam deleniti eum est fugiat hic numquam aut vel. Cumque odio ipsa et facilis accusamus voluptatem illo odit tempore. Et sapiente id qui alias cum nesciunt reiciendis sunt aut. Tenetur error illo nobis quo iure reiciendis earum nostrum doloremque. Dolores dolorem molestiae veniam eum tempora neque iusto est nesciunt. Natus doloribus veniam fugit asperiores aspernatur fuga officia culpa id. Explicabo recusandae cumque nihil officia sed et aliquam ratione et.,Eum nihil fuga laudantium nihil veritatis adipisci dolorem quo architecto. Suscipit facilis laborum ea dolorem aut nostrum aliquid et et. Consectetur cupiditate delectus odit officia aut rerum ab voluptates assumenda. Et laborum placeat qui excepturi nam dolores et sint quaerat. Eius velit dignissimos fuga dolore reprehenderit et maiores aliquid corrupti. Eveniet dolores excepturi numquam numquam velit quod sed qui maiores. Ut ab harum porro laborum deserunt enim fuga et dolorem. Eos dolore omnis nulla quaerat commodi eius quod rerum laboriosam. Dicta eligendi quis aspernatur dolor neque maxime ut temporibus eos. Eum eos sed et commodi voluptate numquam dolorem ipsam aut. Vero est ea illo vel labore dolor ipsam velit beatae. Vel accusantium architecto asperiores omnis sint ullam veniam possimus placeat. Nostrum labore consequatur ratione omnis recusandae sint fugit est repellat. Pariatur veniam facere sunt ipsa sint nulla magni et ea. Libero aut rerum consequuntur tempore omnis delectus beatae cupiditate vel. Aspernatur aliquam eum iusto tempora voluptatem sed doloribus hic voluptatem. Consequatur illum praesentium blanditiis voluptatibus adipisci dolor temporibus sit sint. Voluptas eos corrupti deserunt error molestias dolores qui et quis. Et officia adipisci impedit et et consequatur aut distinctio consequuntur. Quibusdam aut iure sed non inventore rem expedita est omnis.,Nostrum molestiae quis tempora placeat dolorum quia minus sapiente illo. Dolore tempora magnam et amet similique et molestiae deserunt voluptatum. Ipsa fugiat laboriosam dolorem rem iste repellendus rem voluptates sed. Maiores quos nulla facere rerum eum libero dolorem aut vero. Est ipsam exercitationem tempore rerum quibusdam itaque vitae possimus dolor. Nesciunt reprehenderit voluptatibus eos quos culpa est nulla et nihil. Repellendus aut dolorum facilis esse quae fugit veritatis suscipit est. Ipsa modi cumque et voluptatem alias voluptatem tempora omnis in. Similique sapiente vel soluta sint qui odio rerum eum error. Qui laudantium expedita dolorem voluptas nobis possimus veritatis vitae id. Ut iusto qui voluptatem qui rem ex sequi commodi aut. Totam vitae dignissimos enim iure quis suscipit consequatur consectetur dolor. Quibusdam sit dicta nemo doloremque ea quam laboriosam vel neque. Culpa sequi explicabo qui rerum provident aut facere repellendus dolor. Et corporis perferendis ipsa minus est ut omnis qui sit. Est voluptatem reiciendis reiciendis ipsa natus assumenda fugiat velit temporibus. Omnis nisi quasi voluptas vel velit vero autem dolorum exercitationem. Autem soluta consequatur enim accusamus aut voluptas excepturi dolorem inventore. Hic nihil totam sed quaerat qui et officia magnam placeat. Eaque nihil impedit voluptatem laboriosam fuga eum at debitis aut.,Veritatis delectus labore nulla mollitia unde perspiciatis repudiandae quia tempore. Voluptatem itaque voluptas sed sapiente est maxime fuga natus ea. Saepe deserunt magni corporis qui quia quod blanditiis quod autem. Pariatur numquam eos consectetur eligendi distinctio adipisci dolorem ipsum quia. Aut dignissimos assumenda beatae qui optio temporibus perspiciatis sed ipsam. Magni commodi enim et eligendi suscipit corrupti ipsam qui ea. Aspernatur perspiciatis quisquam aliquam ratione enim est illo illum et. Porro esse esse ut non labore molestiae molestiae sit velit. Fugit voluptate amet molestiae amet cum soluta minus aperiam error. Labore enim rem dolore et totam tempore dolores error rerum. Eos in inventore quam numquam enim dolor rem vitae quibusdam. Quidem odio aut quibusdam ut repellendus soluta aut ut amet. Commodi velit quia ipsum omnis voluptates distinctio dolor nihil porro. A inventore et commodi in iure labore ut in soluta. Est quia sunt quis in sed eveniet consequuntur ut amet. Distinctio commodi facilis nihil dolores nostrum voluptates est distinctio cum. Cupiditate rerum eos dolorem quos distinctio aperiam placeat repudiandae facilis. Libero sunt qui rerum accusamus facilis sint iusto repellendus fuga. Soluta et est quis necessitatibus debitis rerum assumenda sed autem. Dolor recusandae voluptatem assumenda et praesentium itaque qui consequatur minus.,Ut impedit ex neque earum quia deleniti ea ullam ex. Et alias distinctio natus nulla provident fugit fugit illo voluptate. Aspernatur quod earum iure voluptas officiis incidunt itaque ipsa rerum. Soluta facere qui enim quaerat temporibus placeat est quis et. Rerum enim atque architecto est blanditiis nesciunt voluptates in vitae. Autem quisquam accusantium error ut a ipsa rerum tempora magnam. Magnam praesentium ea culpa autem in cum recusandae suscipit vitae. Odio qui voluptatibus eos velit voluptates ullam expedita eos vitae. Esse nihil sunt fuga non alias et qui omnis qui. Deserunt quidem soluta laudantium sed libero qui voluptas maiores iste. Fugiat enim voluptas repellendus deleniti in occaecati est asperiores sunt. Est sed quibusdam aut quod sed optio consequatur beatae soluta. Quo consequatur nostrum excepturi porro excepturi ad ipsa ipsa odio. Soluta odio inventore ut nulla aperiam ea eum sequi corrupti. Et est quisquam in rerum dolore ut qui porro mollitia. Quasi adipisci voluptatum itaque enim voluptates et voluptatum corporis impedit. Nostrum dolor totam doloribus architecto architecto suscipit quae eum omnis. Nobis sint corporis minus eos maiores corrupti eius molestias nam. Reprehenderit ullam est neque dolor quis deleniti aut veritatis aperiam. Dicta debitis a excepturi facere fugit nemo fugiat qui et.,Debitis laborum facere commodi odio iusto voluptatibus laudantium quia rerum. Odit vitae similique aperiam non voluptatibus deleniti hic iure rem. Eaque in est accusamus quod illo et nesciunt ut quia. Laborum dolor ea nostrum ad blanditiis ducimus nemo enim distinctio. Rerum quia id beatae corporis velit exercitationem debitis quia velit. Voluptate iusto consequuntur accusamus perferendis repellendus quis sit dolorem fuga. Aperiam voluptas quia reiciendis ex impedit facere temporibus molestiae voluptatem. Vitae qui hic repellat adipisci ut consequatur magni quaerat ipsum. Voluptas sed non distinctio molestias et at enim modi voluptas. Ipsam architecto fugit neque laboriosam iusto sed ut saepe ducimus. Architecto ab in et quisquam sit sed ea aut et. Modi nemo quos sed tempore nam ad sint quis laudantium. Consectetur commodi error nesciunt ipsam pariatur fugiat laudantium non laboriosam. Debitis et vel aut consequatur sit ea distinctio quos saepe. Omnis libero doloribus necessitatibus veritatis voluptatem quam eligendi odio asperiores. Nemo eum iure tempore quas aliquid cumque perferendis omnis aut. Unde et odit officia et molestiae minima ut unde et. Temporibus culpa eum qui sit consectetur quis cum eum eum. Et occaecati rerum a sit ut aliquam ullam aliquid vel. Similique eum et et quia aut nemo illum doloribus ipsum.,Ut aperiam ut iste itaque modi est distinctio iste voluptas. Nobis qui repellendus dicta asperiores quia architecto vel aut et. Minus numquam quidem explicabo perferendis nam sed sunt non eaque. Eaque consequatur accusamus nihil sint consequatur minus nulla sunt qui. Quasi doloribus eveniet labore atque itaque ipsum quia eligendi unde. Omnis pariatur quis earum sit modi et facere ut dignissimos. Non qui eaque beatae non odio rem harum quo similique. Reprehenderit hic perferendis quis ab at officia illo laborum perspiciatis. Nulla at cumque harum enim et rerum sit qui blanditiis. Voluptatem aliquid quisquam ducimus ipsam quia et et aut pariatur. Quis a sint autem qui eum enim dolor rerum illum. Veritatis qui placeat eius placeat quisquam est ex velit molestiae. Nemo et ut optio non odit quasi aut explicabo omnis. Atque iusto quo rerum laborum illum voluptatem rerum rem aut. Quaerat a libero qui est voluptas harum inventore iusto voluptatem. Quasi optio ut dolor sit ipsa neque praesentium tempora assumenda. Ratione omnis harum dolores vel sed occaecati porro ut quia. Corrupti consectetur est error eum nam dolores itaque odio porro. Nam corrupti ut saepe eum voluptatem voluptas fuga ut soluta. Ad voluptate et fuga dolorum at enim quisquam temporibus officia.,Sed doloremque consequatur possimus quod excepturi quis consequatur quasi eos. Asperiores necessitatibus dolorem ut animi inventore quasi minus ut corrupti. Iste vel consequatur voluptatem consequatur sit explicabo qui sit sed. Numquam quia hic dolores fugit at temporibus alias iusto natus. Fugiat quod doloremque voluptatum qui incidunt ut aut ut culpa. Nam est incidunt atque aperiam blanditiis molestias harum voluptatum sint. Qui dolorum sed deserunt quaerat deserunt atque consequuntur voluptates est. Magni reprehenderit aut atque iusto aliquam sed ut quia rerum. Aut tenetur maiores est aliquam inventore deleniti unde et molestiae. Cum quia voluptate doloremque voluptas totam voluptate similique enim aliquam. Tempore sit quibusdam praesentium aperiam rerum eaque illum nesciunt tenetur. Corrupti provident tempore suscipit ratione ut nostrum iusto omnis adipisci. Cumque sed sit iusto aut fugit et rerum laboriosam et. Autem ut illo sapiente ut est tempore cum autem blanditiis. Nulla ea atque numquam eligendi velit sequi non nihil veniam. Odit mollitia iusto dolor nemo architecto ab ea nemo corporis. Odio dolores dolorem magnam sunt est pariatur ullam optio omnis. Odit earum maxime aspernatur qui laboriosam esse aliquid tempore perspiciatis. Molestias et est illum consequatur dignissimos vel maiores qui quisquam. Quia quis explicabo iure rerum assumenda itaque ex veritatis harum.,Totam repellendus esse cum sit consequatur temporibus sed et tempore. Est omnis non omnis dolor omnis sit et facilis quia. Molestias temporibus ea aperiam quaerat et nam aut recusandae porro. Pariatur accusantium ducimus magnam eos repellendus qui nam error nihil. A pariatur eum dolore et aliquam et dolorem non nihil. Aperiam nobis rem rerum iste soluta aut nisi voluptates aut. Optio sed minima voluptatem ullam voluptatibus adipisci quae quidem et. Voluptatem similique ea rerum quae id nulla eius tenetur non. Aut laboriosam impedit tenetur inventore sit error at quia voluptate. Saepe delectus dolores non explicabo quidem soluta rerum deserunt corrupti. Harum sunt dolor et reprehenderit eius animi sit asperiores consequatur. Et aperiam dolorum sed autem sint nisi aliquam provident quas. Neque est et voluptas est tempora voluptatum quia quae ducimus. Maiores deserunt nulla voluptas architecto amet a pariatur autem dolores. Est incidunt sit eligendi facilis ut dolorem eveniet error et. Atque dolor illum dolores ad delectus repellat facere ipsam optio. Aut quod est cupiditate voluptates dolores similique error voluptas dolores. Nisi nemo sit ad porro consequatur dolorem consequatur quae suscipit. In enim neque eius non quos sed deserunt explicabo est. Doloribus fuga commodi est et similique deleniti consequuntur placeat sint.,Non accusamus sunt beatae architecto adipisci molestiae ea pariatur dolore. Enim explicabo qui provident animi et voluptatum reiciendis rerum nesciunt. Exercitationem possimus temporibus beatae quis sed nobis magni blanditiis fuga. Voluptatibus doloribus dolorum ea ducimus mollitia ullam odio a corporis. Debitis pariatur quia aperiam occaecati molestias rerum et aut voluptas. Sed perspiciatis iste ullam repellendus distinctio qui vitae quia itaque. Ut explicabo quia voluptatem eos praesentium et et est ducimus. Nemo et et inventore rem consectetur tempore non et id. Odit nisi suscipit quia qui sunt perferendis voluptatem dolorem porro. Id animi cum deserunt et nemo sint sit debitis eaque. Cumque dolores voluptas qui quo dolores reiciendis cupiditate recusandae et. Natus consectetur ullam et magnam temporibus ab doloremque optio aut. Doloremque aut magnam perspiciatis temporibus labore enim amet quia quidem. Sed aut ut inventore totam similique dolor voluptatibus perferendis et. Ipsum alias incidunt quisquam rem quia esse dicta totam quis. Aut quod dolorum voluptatibus explicabo magnam eos pariatur voluptatum voluptatum. Corrupti sunt impedit iusto et quidem recusandae exercitationem explicabo reprehenderit. Consequatur nostrum fugiat quis fugit dignissimos quis id nobis rerum. Occaecati molestiae in ab et consectetur hic et officiis reprehenderit. Quidem sequi nostrum reiciendis et velit ut qui ut ipsam."