read excel sheet by name to json

This is a function to read out the content out of a excel sheet to json array

read_sheet_by_name(file_path, sheet_name)

Examples

read_sheet_by_name("example_input.xlsx","Test1")
#> [[1]]
#> [[1]]$drat
#> [1] "3.9"
#> 
#> [[1]]$`_row`
#> [1] "Mazda RX4"
#> 
#> [[1]]$vs
#> [1] "0"
#> 
#> [[1]]$disp
#> [1] "160"
#> 
#> [[1]]$qsec
#> [1] "16.46"
#> 
#> [[1]]$am
#> [1] "1"
#> 
#> [[1]]$carb
#> [1] "4"
#> 
#> [[1]]$cyl
#> [1] "6"
#> 
#> [[1]]$mpg
#> [1] "21"
#> 
#> [[1]]$wt
#> [1] "2.62"
#> 
#> [[1]]$gear
#> [1] "4"
#> 
#> [[1]]$hp
#> [1] "110"
#> 
#> 
#> [[2]]
#> [[2]]$cyl
#> [1] "6"
#> 
#> [[2]]$gear
#> [1] "4"
#> 
#> [[2]]$carb
#> [1] "4"
#> 
#> [[2]]$am
#> [1] "1"
#> 
#> [[2]]$mpg
#> [1] "21"
#> 
#> [[2]]$drat
#> [1] "3.9"
#> 
#> [[2]]$qsec
#> [1] "17.02"
#> 
#> [[2]]$hp
#> [1] "110"
#> 
#> [[2]]$disp
#> [1] "160"
#> 
#> [[2]]$vs
#> [1] "0"
#> 
#> [[2]]$wt
#> [1] "2.875"
#> 
#> [[2]]$`_row`
#> [1] "Mazda RX4 Wag"
#> 
#> 
#> [[3]]
#> [[3]]$drat
#> [1] "3.85"
#> 
#> [[3]]$qsec
#> [1] "18.61"
#> 
#> [[3]]$disp
#> [1] "108"
#> 
#> [[3]]$wt
#> [1] "2.32"
#> 
#> [[3]]$vs
#> [1] "1"
#> 
#> [[3]]$am
#> [1] "1"
#> 
#> [[3]]$cyl
#> [1] "4"
#> 
#> [[3]]$carb
#> [1] "1"
#> 
#> [[3]]$`_row`
#> [1] "Datsun 710"
#> 
#> [[3]]$gear
#> [1] "4"
#> 
#> [[3]]$hp
#> [1] "93"
#> 
#> [[3]]$mpg
#> [1] "22.8"
#> 
#> 
#> [[4]]
#> [[4]]$am
#> [1] "0"
#> 
#> [[4]]$`_row`
#> [1] "Hornet 4 Drive"
#> 
#> [[4]]$carb
#> [1] "1"
#> 
#> [[4]]$disp
#> [1] "258"
#> 
#> [[4]]$gear
#> [1] "3"
#> 
#> [[4]]$hp
#> [1] "110"
#> 
#> [[4]]$mpg
#> [1] "21.4"
#> 
#> [[4]]$qsec
#> [1] "19.44"
#> 
#> [[4]]$wt
#> [1] "3.215"
#> 
#> [[4]]$drat
#> [1] "3.08"
#> 
#> [[4]]$cyl
#> [1] "6"
#> 
#> [[4]]$vs
#> [1] "1"
#> 
#> 
#> [[5]]
#> [[5]]$am
#> [1] "0"
#> 
#> [[5]]$cyl
#> [1] "8"
#> 
#> [[5]]$drat
#> [1] "3.15"
#> 
#> [[5]]$wt
#> [1] "3.44"
#> 
#> [[5]]$carb
#> [1] "2"
#> 
#> [[5]]$gear
#> [1] "3"
#> 
#> [[5]]$hp
#> [1] "175"
#> 
#> [[5]]$`_row`
#> [1] "Hornet Sportabout"
#> 
#> [[5]]$disp
#> [1] "360"
#> 
#> [[5]]$mpg
#> [1] "18.7"
#> 
#> [[5]]$qsec
#> [1] "17.02"
#> 
#> [[5]]$vs
#> [1] "0"
#> 
#> 
#> [[6]]
#> [[6]]$gear
#> [1] "3"
#> 
#> [[6]]$disp
#> [1] "225"
#> 
#> [[6]]$hp
#> [1] "105"
#> 
#> [[6]]$qsec
#> [1] "20.22"
#> 
#> [[6]]$wt
#> [1] "3.46"
#> 
#> [[6]]$vs
#> [1] "1"
#> 
#> [[6]]$cyl
#> [1] "6"
#> 
#> [[6]]$drat
#> [1] "2.76"
#> 
#> [[6]]$am
#> [1] "0"
#> 
#> [[6]]$mpg
#> [1] "18.1"
#> 
#> [[6]]$`_row`
#> [1] "Valiant"
#> 
#> [[6]]$carb
#> [1] "1"
#> 
#> 
#> [[7]]
#> [[7]]$drat
#> [1] "3.21"
#> 
#> [[7]]$qsec
#> [1] "15.84"
#> 
#> [[7]]$carb
#> [1] "4"
#> 
#> [[7]]$wt
#> [1] "3.57"
#> 
#> [[7]]$cyl
#> [1] "8"
#> 
#> [[7]]$am
#> [1] "0"
#> 
#> [[7]]$gear
#> [1] "3"
#> 
#> [[7]]$hp
#> [1] "245"
#> 
#> [[7]]$mpg
#> [1] "14.3"
#> 
#> [[7]]$vs
#> [1] "0"
#> 
#> [[7]]$disp
#> [1] "360"
#> 
#> [[7]]$`_row`
#> [1] "Duster 360"
#> 
#> 
#> [[8]]
#> [[8]]$drat
#> [1] "3.69"
#> 
#> [[8]]$`_row`
#> [1] "Merc 240D"
#> 
#> [[8]]$vs
#> [1] "1"
#> 
#> [[8]]$wt
#> [1] "3.19"
#> 
#> [[8]]$carb
#> [1] "2"
#> 
#> [[8]]$cyl
#> [1] "4"
#> 
#> [[8]]$am
#> [1] "0"
#> 
#> [[8]]$gear
#> [1] "4"
#> 
#> [[8]]$disp
#> [1] "146.7"
#> 
#> [[8]]$hp
#> [1] "62"
#> 
#> [[8]]$mpg
#> [1] "24.4"
#> 
#> [[8]]$qsec
#> [1] "20"
#> 
#> 
#> [[9]]
#> [[9]]$am
#> [1] "0"
#> 
#> [[9]]$drat
#> [1] "3.92"
#> 
#> [[9]]$`_row`
#> [1] "Merc 230"
#> 
#> [[9]]$carb
#> [1] "2"
#> 
#> [[9]]$gear
#> [1] "4"
#> 
#> [[9]]$hp
#> [1] "95"
#> 
#> [[9]]$qsec
#> [1] "22.9"
#> 
#> [[9]]$wt
#> [1] "3.15"
#> 
#> [[9]]$mpg
#> [1] "22.8"
#> 
#> [[9]]$cyl
#> [1] "4"
#> 
#> [[9]]$disp
#> [1] "140.8"
#> 
#> [[9]]$vs
#> [1] "1"
#> 
#> 
#> [[10]]
#> [[10]]$carb
#> [1] "4"
#> 
#> [[10]]$hp
#> [1] "123"
#> 
#> [[10]]$`_row`
#> [1] "Merc 280"
#> 
#> [[10]]$wt
#> [1] "3.44"
#> 
#> [[10]]$cyl
#> [1] "6"
#> 
#> [[10]]$disp
#> [1] "167.6"
#> 
#> [[10]]$am
#> [1] "0"
#> 
#> [[10]]$drat
#> [1] "3.92"
#> 
#> [[10]]$gear
#> [1] "4"
#> 
#> [[10]]$mpg
#> [1] "19.2"
#> 
#> [[10]]$qsec
#> [1] "18.3"
#> 
#> [[10]]$vs
#> [1] "1"
#> 
#> 
#> [[11]]
#> [[11]]$`_row`
#> [1] "Merc 280C"
#> 
#> [[11]]$mpg
#> [1] "17.8"
#> 
#> [[11]]$carb
#> [1] "4"
#> 
#> [[11]]$vs
#> [1] "1"
#> 
#> [[11]]$am
#> [1] "0"
#> 
#> [[11]]$wt
#> [1] "3.44"
#> 
#> [[11]]$disp
#> [1] "167.6"
#> 
#> [[11]]$hp
#> [1] "123"
#> 
#> [[11]]$gear
#> [1] "4"
#> 
#> [[11]]$drat
#> [1] "3.92"
#> 
#> [[11]]$qsec
#> [1] "18.9"
#> 
#> [[11]]$cyl
#> [1] "6"
#> 
#> 
#> [[12]]
#> [[12]]$gear
#> [1] "3"
#> 
#> [[12]]$cyl
#> [1] "8"
#> 
#> [[12]]$am
#> [1] "0"
#> 
#> [[12]]$drat
#> [1] "3.07"
#> 
#> [[12]]$disp
#> [1] "275.8"
#> 
#> [[12]]$hp
#> [1] "180"
#> 
#> [[12]]$mpg
#> [1] "16.4"
#> 
#> [[12]]$`_row`
#> [1] "Merc 450SE"
#> 
#> [[12]]$qsec
#> [1] "17.4"
#> 
#> [[12]]$wt
#> [1] "4.07"
#> 
#> [[12]]$carb
#> [1] "3"
#> 
#> [[12]]$vs
#> [1] "0"
#> 
#> 
#> [[13]]
#> [[13]]$drat
#> [1] "3.07"
#> 
#> [[13]]$gear
#> [1] "3"
#> 
#> [[13]]$disp
#> [1] "275.8"
#> 
#> [[13]]$cyl
#> [1] "8"
#> 
#> [[13]]$carb
#> [1] "3"
#> 
#> [[13]]$hp
#> [1] "180"
#> 
#> [[13]]$mpg
#> [1] "17.3"
#> 
#> [[13]]$qsec
#> [1] "17.6"
#> 
#> [[13]]$vs
#> [1] "0"
#> 
#> [[13]]$wt
#> [1] "3.73"
#> 
#> [[13]]$`_row`
#> [1] "Merc 450SL"
#> 
#> [[13]]$am
#> [1] "0"
#> 
#> 
#> [[14]]
#> [[14]]$disp
#> [1] "275.8"
#> 
#> [[14]]$`_row`
#> [1] "Merc 450SLC"
#> 
#> [[14]]$hp
#> [1] "180"
#> 
#> [[14]]$am
#> [1] "0"
#> 
#> [[14]]$drat
#> [1] "3.07"
#> 
#> [[14]]$gear
#> [1] "3"
#> 
#> [[14]]$mpg
#> [1] "15.2"
#> 
#> [[14]]$vs
#> [1] "0"
#> 
#> [[14]]$wt
#> [1] "3.78"
#> 
#> [[14]]$carb
#> [1] "3"
#> 
#> [[14]]$qsec
#> [1] "18"
#> 
#> [[14]]$cyl
#> [1] "8"
#> 
#> 
#> [[15]]
#> [[15]]$carb
#> [1] "4"
#> 
#> [[15]]$mpg
#> [1] "10.4"
#> 
#> [[15]]$`_row`
#> [1] "Cadillac Fleetwood"
#> 
#> [[15]]$gear
#> [1] "3"
#> 
#> [[15]]$vs
#> [1] "0"
#> 
#> [[15]]$wt
#> [1] "5.25"
#> 
#> [[15]]$qsec
#> [1] "17.98"
#> 
#> [[15]]$drat
#> [1] "2.93"
#> 
#> [[15]]$hp
#> [1] "205"
#> 
#> [[15]]$am
#> [1] "0"
#> 
#> [[15]]$disp
#> [1] "472"
#> 
#> [[15]]$cyl
#> [1] "8"
#> 
#> 
#> [[16]]
#> [[16]]$disp
#> [1] "460"
#> 
#> [[16]]$carb
#> [1] "4"
#> 
#> [[16]]$qsec
#> [1] "17.82"
#> 
#> [[16]]$`_row`
#> [1] "Lincoln Continental"
#> 
#> [[16]]$cyl
#> [1] "8"
#> 
#> [[16]]$am
#> [1] "0"
#> 
#> [[16]]$drat
#> [1] "3"
#> 
#> [[16]]$gear
#> [1] "3"
#> 
#> [[16]]$hp
#> [1] "215"
#> 
#> [[16]]$mpg
#> [1] "10.4"
#> 
#> [[16]]$vs
#> [1] "0"
#> 
#> [[16]]$wt
#> [1] "5.424"
#> 
#> 
#> [[17]]
#> [[17]]$wt
#> [1] "5.345"
#> 
#> [[17]]$carb
#> [1] "4"
#> 
#> [[17]]$mpg
#> [1] "14.7"
#> 
#> [[17]]$gear
#> [1] "3"
#> 
#> [[17]]$qsec
#> [1] "17.42"
#> 
#> [[17]]$cyl
#> [1] "8"
#> 
#> [[17]]$`_row`
#> [1] "Chrysler Imperial"
#> 
#> [[17]]$am
#> [1] "0"
#> 
#> [[17]]$drat
#> [1] "3.23"
#> 
#> [[17]]$hp
#> [1] "230"
#> 
#> [[17]]$vs
#> [1] "0"
#> 
#> [[17]]$disp
#> [1] "440"
#> 
#> 
#> [[18]]
#> [[18]]$cyl
#> [1] "4"
#> 
#> [[18]]$qsec
#> [1] "19.47"
#> 
#> [[18]]$vs
#> [1] "1"
#> 
#> [[18]]$am
#> [1] "1"
#> 
#> [[18]]$gear
#> [1] "4"
#> 
#> [[18]]$carb
#> [1] "1"
#> 
#> [[18]]$wt
#> [1] "2.2"
#> 
#> [[18]]$mpg
#> [1] "32.4"
#> 
#> [[18]]$disp
#> [1] "78.7"
#> 
#> [[18]]$drat
#> [1] "4.08"
#> 
#> [[18]]$hp
#> [1] "66"
#> 
#> [[18]]$`_row`
#> [1] "Fiat 128"
#> 
#> 
#> [[19]]
#> [[19]]$gear
#> [1] "4"
#> 
#> [[19]]$am
#> [1] "1"
#> 
#> [[19]]$mpg
#> [1] "30.4"
#> 
#> [[19]]$disp
#> [1] "75.7"
#> 
#> [[19]]$qsec
#> [1] "18.52"
#> 
#> [[19]]$vs
#> [1] "1"
#> 
#> [[19]]$hp
#> [1] "52"
#> 
#> [[19]]$drat
#> [1] "4.93"
#> 
#> [[19]]$wt
#> [1] "1.615"
#> 
#> [[19]]$carb
#> [1] "2"
#> 
#> [[19]]$`_row`
#> [1] "Honda Civic"
#> 
#> [[19]]$cyl
#> [1] "4"
#> 
#> 
#> [[20]]
#> [[20]]$vs
#> [1] "1"
#> 
#> [[20]]$am
#> [1] "1"
#> 
#> [[20]]$`_row`
#> [1] "Toyota Corolla"
#> 
#> [[20]]$carb
#> [1] "1"
#> 
#> [[20]]$cyl
#> [1] "4"
#> 
#> [[20]]$wt
#> [1] "1.835"
#> 
#> [[20]]$disp
#> [1] "71.1"
#> 
#> [[20]]$gear
#> [1] "4"
#> 
#> [[20]]$drat
#> [1] "4.22"
#> 
#> [[20]]$hp
#> [1] "65"
#> 
#> [[20]]$mpg
#> [1] "33.9"
#> 
#> [[20]]$qsec
#> [1] "19.9"
#> 
#> 
#> [[21]]
#> [[21]]$am
#> [1] "0"
#> 
#> [[21]]$carb
#> [1] "1"
#> 
#> [[21]]$cyl
#> [1] "4"
#> 
#> [[21]]$vs
#> [1] "1"
#> 
#> [[21]]$hp
#> [1] "97"
#> 
#> [[21]]$drat
#> [1] "3.7"
#> 
#> [[21]]$`_row`
#> [1] "Toyota Corona"
#> 
#> [[21]]$gear
#> [1] "3"
#> 
#> [[21]]$mpg
#> [1] "21.5"
#> 
#> [[21]]$qsec
#> [1] "20.01"
#> 
#> [[21]]$disp
#> [1] "120.1"
#> 
#> [[21]]$wt
#> [1] "2.465"
#> 
#> 
#> [[22]]
#> [[22]]$disp
#> [1] "318"
#> 
#> [[22]]$`_row`
#> [1] "Dodge Challenger"
#> 
#> [[22]]$am
#> [1] "0"
#> 
#> [[22]]$gear
#> [1] "3"
#> 
#> [[22]]$vs
#> [1] "0"
#> 
#> [[22]]$wt
#> [1] "3.52"
#> 
#> [[22]]$drat
#> [1] "2.76"
#> 
#> [[22]]$mpg
#> [1] "15.5"
#> 
#> [[22]]$cyl
#> [1] "8"
#> 
#> [[22]]$qsec
#> [1] "16.87"
#> 
#> [[22]]$carb
#> [1] "2"
#> 
#> [[22]]$hp
#> [1] "150"
#> 
#> 
#> [[23]]
#> [[23]]$disp
#> [1] "304"
#> 
#> [[23]]$cyl
#> [1] "8"
#> 
#> [[23]]$`_row`
#> [1] "AMC Javelin"
#> 
#> [[23]]$am
#> [1] "0"
#> 
#> [[23]]$gear
#> [1] "3"
#> 
#> [[23]]$vs
#> [1] "0"
#> 
#> [[23]]$mpg
#> [1] "15.2"
#> 
#> [[23]]$wt
#> [1] "3.435"
#> 
#> [[23]]$carb
#> [1] "2"
#> 
#> [[23]]$qsec
#> [1] "17.3"
#> 
#> [[23]]$drat
#> [1] "3.15"
#> 
#> [[23]]$hp
#> [1] "150"
#> 
#> 
#> [[24]]
#> [[24]]$disp
#> [1] "350"
#> 
#> [[24]]$`_row`
#> [1] "Camaro Z28"
#> 
#> [[24]]$gear
#> [1] "3"
#> 
#> [[24]]$carb
#> [1] "4"
#> 
#> [[24]]$drat
#> [1] "3.73"
#> 
#> [[24]]$hp
#> [1] "245"
#> 
#> [[24]]$mpg
#> [1] "13.3"
#> 
#> [[24]]$am
#> [1] "0"
#> 
#> [[24]]$qsec
#> [1] "15.41"
#> 
#> [[24]]$vs
#> [1] "0"
#> 
#> [[24]]$cyl
#> [1] "8"
#> 
#> [[24]]$wt
#> [1] "3.84"
#> 
#> 
#> [[25]]
#> [[25]]$`_row`
#> [1] "Pontiac Firebird"
#> 
#> [[25]]$carb
#> [1] "2"
#> 
#> [[25]]$am
#> [1] "0"
#> 
#> [[25]]$drat
#> [1] "3.08"
#> 
#> [[25]]$gear
#> [1] "3"
#> 
#> [[25]]$disp
#> [1] "400"
#> 
#> [[25]]$vs
#> [1] "0"
#> 
#> [[25]]$wt
#> [1] "3.845"
#> 
#> [[25]]$mpg
#> [1] "19.2"
#> 
#> [[25]]$qsec
#> [1] "17.05"
#> 
#> [[25]]$cyl
#> [1] "8"
#> 
#> [[25]]$hp
#> [1] "175"
#> 
#> 
#> [[26]]
#> [[26]]$disp
#> [1] "79"
#> 
#> [[26]]$gear
#> [1] "4"
#> 
#> [[26]]$am
#> [1] "1"
#> 
#> [[26]]$wt
#> [1] "1.935"
#> 
#> [[26]]$hp
#> [1] "66"
#> 
#> [[26]]$qsec
#> [1] "18.9"
#> 
#> [[26]]$cyl
#> [1] "4"
#> 
#> [[26]]$carb
#> [1] "1"
#> 
#> [[26]]$mpg
#> [1] "27.3"
#> 
#> [[26]]$`_row`
#> [1] "Fiat X1-9"
#> 
#> [[26]]$drat
#> [1] "4.08"
#> 
#> [[26]]$vs
#> [1] "1"
#> 
#> 
#> [[27]]
#> [[27]]$am
#> [1] "1"
#> 
#> [[27]]$wt
#> [1] "2.14"
#> 
#> [[27]]$carb
#> [1] "2"
#> 
#> [[27]]$hp
#> [1] "91"
#> 
#> [[27]]$`_row`
#> [1] "Porsche 914-2"
#> 
#> [[27]]$mpg
#> [1] "26"
#> 
#> [[27]]$vs
#> [1] "0"
#> 
#> [[27]]$qsec
#> [1] "16.7"
#> 
#> [[27]]$cyl
#> [1] "4"
#> 
#> [[27]]$disp
#> [1] "120.3"
#> 
#> [[27]]$drat
#> [1] "4.43"
#> 
#> [[27]]$gear
#> [1] "5"
#> 
#> 
#> [[28]]
#> [[28]]$am
#> [1] "1"
#> 
#> [[28]]$mpg
#> [1] "30.4"
#> 
#> [[28]]$vs
#> [1] "1"
#> 
#> [[28]]$gear
#> [1] "5"
#> 
#> [[28]]$cyl
#> [1] "4"
#> 
#> [[28]]$disp
#> [1] "95.1"
#> 
#> [[28]]$drat
#> [1] "3.77"
#> 
#> [[28]]$carb
#> [1] "2"
#> 
#> [[28]]$qsec
#> [1] "16.9"
#> 
#> [[28]]$hp
#> [1] "113"
#> 
#> [[28]]$`_row`
#> [1] "Lotus Europa"
#> 
#> [[28]]$wt
#> [1] "1.513"
#> 
#> 
#> [[29]]
#> [[29]]$gear
#> [1] "5"
#> 
#> [[29]]$wt
#> [1] "3.17"
#> 
#> [[29]]$drat
#> [1] "4.22"
#> 
#> [[29]]$`_row`
#> [1] "Ford Pantera L"
#> 
#> [[29]]$carb
#> [1] "4"
#> 
#> [[29]]$disp
#> [1] "351"
#> 
#> [[29]]$cyl
#> [1] "8"
#> 
#> [[29]]$hp
#> [1] "264"
#> 
#> [[29]]$am
#> [1] "1"
#> 
#> [[29]]$mpg
#> [1] "15.8"
#> 
#> [[29]]$qsec
#> [1] "14.5"
#> 
#> [[29]]$vs
#> [1] "0"
#> 
#> 
#> [[30]]
#> [[30]]$`_row`
#> [1] "Ferrari Dino"
#> 
#> [[30]]$carb
#> [1] "6"
#> 
#> [[30]]$disp
#> [1] "145"
#> 
#> [[30]]$drat
#> [1] "3.62"
#> 
#> [[30]]$gear
#> [1] "5"
#> 
#> [[30]]$mpg
#> [1] "19.7"
#> 
#> [[30]]$hp
#> [1] "175"
#> 
#> [[30]]$vs
#> [1] "0"
#> 
#> [[30]]$am
#> [1] "1"
#> 
#> [[30]]$qsec
#> [1] "15.5"
#> 
#> [[30]]$cyl
#> [1] "6"
#> 
#> [[30]]$wt
#> [1] "2.77"
#> 
#> 
#> [[31]]
#> [[31]]$vs
#> [1] "0"
#> 
#> [[31]]$drat
#> [1] "3.54"
#> 
#> [[31]]$am
#> [1] "1"
#> 
#> [[31]]$`_row`
#> [1] "Maserati Bora"
#> 
#> [[31]]$hp
#> [1] "335"
#> 
#> [[31]]$disp
#> [1] "301"
#> 
#> [[31]]$cyl
#> [1] "8"
#> 
#> [[31]]$qsec
#> [1] "14.6"
#> 
#> [[31]]$gear
#> [1] "5"
#> 
#> [[31]]$wt
#> [1] "3.57"
#> 
#> [[31]]$carb
#> [1] "8"
#> 
#> [[31]]$mpg
#> [1] "15"
#> 
#> 
#> [[32]]
#> [[32]]$hp
#> [1] "109"
#> 
#> [[32]]$drat
#> [1] "4.11"
#> 
#> [[32]]$vs
#> [1] "1"
#> 
#> [[32]]$mpg
#> [1] "21.4"
#> 
#> [[32]]$qsec
#> [1] "18.6"
#> 
#> [[32]]$cyl
#> [1] "4"
#> 
#> [[32]]$wt
#> [1] "2.78"
#> 
#> [[32]]$disp
#> [1] "121"
#> 
#> [[32]]$`_row`
#> [1] "Volvo 142E"
#> 
#> [[32]]$gear
#> [1] "4"
#> 
#> [[32]]$carb
#> [1] "2"
#> 
#> [[32]]$am
#> [1] "1"
#> 
#>